Move post-saturation search into the runtime
Core now stops at egglog saturation: `Graph::build_search_space` produces a
`SearchSpace` (one saturated e-graph per dynamic-dim bucket combination, the
registered ops, and custom ops resolved to LLIR), and `Runtime::compile`
owns everything after it — search strategy, profiling, and loading. Every
profile/filter/bucket hook is removed from the `Runtime` trait.

The genetic search Luminal has always used moves into `luminal::search` as
utilities a runtime may use, compose, or ignore:

- `extract_one`: one valid program, for runtimes with nothing to rank on.
- `GeneticSearch`: the GA as a pull-style state machine (`next_candidate` /
  `report`) that owns generations, mutation, stagnation restarts, dedup,
  budgets, the early-stop hint, and progress output.
- `Finalists` and `BucketLattice`: lazy hard-filtered finalist extraction
  and best-first bucket-set selection under an aggregate constraint.
- `genetic_search`: closure sugar composing the three into the stock flow.

Runtimes: the reference runtime extracts one program and never profiles;
Metal calls the sugar; CUDA drives the state machine explicitly in
`luminal_cuda_lite/src/search.rs`, compiling, validating, installing and
profiling each candidate in one step and installing the aggregate-validated
bucket set without recompiling it at load.

Loop unrolling, LLIR dumps/logs, and the packed rolled graph move out of
`graph.rs` into `search::{unroll, diagnostics, packed}`. `spec.md` and
`AGENTS.md` describe the new boundary.

Verified in-process against a verbatim copy of the previous driver: on the
same `SearchSpace` and seed, the new utilities reproduce its trajectory
exactly (900 trace lines over 5 seeds x bucketed/unbucketed x stagnation
restart on/off). One designed difference: the unified lattice validates the
single-bucket set too, where the old unbucketed path skipped aggregate
validation.
