+++ src/graph.rs
@@ -1,7 +1,8 @@
 use crate::egglog_utils::{
-    IndexedChoiceSet, LlirExtractor, count_choice_sets_up_to, hlir_to_egglog, log_channel_enabled,
-    run_egglog_with_late_passes_interval_analysis_and_log,
+    hlir_to_egglog, log_channel_enabled, run_egglog_with_late_passes_interval_analysis_and_log,
 };
+pub use crate::search::unroll::{collapse_loops_to_first_iter, unroll_loops_in_llir};
+use crate::search::{BucketSearchSpace, SearchSpace, bucket_index_combinations};
 use crate::shape::{DimInterval, DynDimIntervals};
 use crate::{
     egglog_utils::SerializedEGraph,
@@ -10,19 +11,10 @@ use crate::{
 use crate::{hlir::CustomOpKind, op::*, prelude::*};
 use colored::Colorize;
 use itertools::Itertools;
-use petgraph::{
-    Direction,
-    dot::{Config, Dot},
-    stable_graph::StableGraph,
-    visit::EdgeRef,
-};
+use petgraph::{Direction, stable_graph::StableGraph, visit::EdgeRef};
 use rustc_hash::{FxHashMap, FxHashSet};
 use std::{
-    any::TypeId,
-    collections::hash_map::{DefaultHasher, Entry},
-    fmt::{Debug, Write as FmtWrite},
-    hash::{Hash, Hasher},
-    io::Write,
+    fmt::Debug,
     ops::{Deref, DerefMut},
     sync::Arc,
 };
@@ -30,97 +22,6 @@ use tracing;
 
 pub type LLIRGraph = StableGraph<LLIROp, ()>;
 pub type HLIRGraph = StableGraph<Box<dyn HLIROp>, ()>;
-
-/// Dense, internal form used between indexed e-graph extraction and loop
-/// materialization. Runtimes continue to receive [`LLIRGraph`]; this only
-/// avoids constructing a throwaway rolled `StableGraph` for every search
-/// candidate.
-pub(crate) struct PackedLLIRGraph {
-    pub(crate) ops: Vec<LLIROp>,
-    pub(crate) incoming_offsets: Vec<usize>,
-    pub(crate) incoming_sources: Vec<usize>,
-    pub(crate) outgoing_offsets: Vec<usize>,
-    pub(crate) outgoing_targets: Vec<usize>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-struct LlirFingerprint(u64, u64);
-
-struct LlirFingerprintHasher {
-    first: DefaultHasher,
-    second: DefaultHasher,
-}
-
-impl LlirFingerprintHasher {
-    fn new() -> Self {
-        let mut first = DefaultHasher::new();
-        let mut second = DefaultHasher::new();
-        0x243f_6a88_85a3_08d3_u64.hash(&mut first);
-        0x1319_8a2e_0370_7344_u64.hash(&mut second);
-        Self { first, second }
-    }
-
-    fn fingerprint(&self) -> LlirFingerprint {
-        LlirFingerprint(self.first.finish(), self.second.finish())
-    }
-}
-
-impl Hasher for LlirFingerprintHasher {
-    fn finish(&self) -> u64 {
-        self.first.finish()
-    }
-
-    fn write(&mut self, bytes: &[u8]) {
-        self.first.write(bytes);
-        self.second.write(bytes);
-    }
-}
-
-struct HashWriter<'a>(&'a mut LlirFingerprintHasher);
-
-impl FmtWrite for HashWriter<'_> {
-    fn write_str(&mut self, value: &str) -> std::fmt::Result {
-        self.0.write(value.as_bytes());
-        Ok(())
-    }
-}
-
-impl PackedLLIRGraph {
-    /// Fingerprint the exact LLIR program represented by this packed graph.
-    /// Packed extraction has deterministic node ordering, so the op sequence
-    /// plus ordered incoming edges completely describes the graph; outgoing
-    /// edges are derived from those inputs and do not need to be hashed again.
-    fn fingerprint(&self) -> LlirFingerprint {
-        let mut hasher = LlirFingerprintHasher::new();
-        self.ops.len().hash(&mut hasher);
-        for op in &self.ops {
-            // Debug is the full semantic representation of a type-erased LLIR
-            // op. Display is intentionally unsuitable: many ops omit their
-            // shape and stride metadata there. Stream directly into the hash
-            // to avoid allocating one String per op and candidate.
-            write!(HashWriter(&mut hasher), "{op:?}").unwrap();
-            hasher.write_u8(0xff);
-        }
-        self.incoming_offsets.hash(&mut hasher);
-        self.incoming_sources.hash(&mut hasher);
-        hasher.fingerprint()
-    }
-
-    pub(crate) fn to_stable(&self) -> LLIRGraph {
-        let mut graph = LLIRGraph::with_capacity(self.ops.len(), self.incoming_sources.len());
-        for op in &self.ops {
-            graph.add_node(op.clone());
-        }
-        for destination in 0..graph.node_count() {
-            for &source in &self.incoming_sources
-                [self.incoming_offsets[destination]..self.incoming_offsets[destination + 1]]
-            {
-                graph.add_edge(NodeIndex::new(source), NodeIndex::new(destination), ());
-            }
-        }
-        graph
-    }
-}
 
 #[derive(Debug, Clone)]
 struct RollingOccurrence {
@@ -166,54 +67,6 @@ struct RollingRejectedCandidate {
 struct RollingSearchReport {
     candidate: Option<RollingCandidate>,
     diagnostics: RollingSearchDiagnostics,
-}
-
-#[derive(Debug, Clone, Default)]
-struct SearchSpaceContext {
-    bucket_indices: DynMap,
-    representative_dyn_map: DynMap,
-    intervals: DynDimIntervals,
-}
-
-#[derive(Debug, Clone)]
-struct SearchProfileBucketContext {
-    dim_buckets: FxHashMap<Symbol, Vec<DimBucket>>,
-    bucket_indices: DynMap,
-    representative_dyn_map: DynMap,
-}
-
-struct Finalist<M> {
-    metric: M,
-    pre_unroll: Option<LLIRGraph>,
-    llir: LLIRGraph,
-}
-
-struct LazyFinalists<M> {
-    ranked: Vec<(M, IndexedChoiceSet)>,
-    next_ranked: usize,
-    finalists: Vec<Finalist<M>>,
-    rejections: usize,
-    last_rejection: Option<String>,
-    stopped_reason: Option<String>,
-}
-
-impl<M> LazyFinalists<M> {
-    fn new(ranked: Vec<(M, IndexedChoiceSet)>) -> Self {
-        Self {
-            ranked,
-            next_ranked: 0,
-            finalists: Vec::new(),
-            rejections: 0,
-            last_rejection: None,
-            stopped_reason: None,
-        }
-    }
-}
-
-struct BucketFinalistSearch<M> {
-    context: SearchProfileBucketContext,
-    egraph_index: usize,
-    candidates: LazyFinalists<M>,
 }
 
 /// A compiled bucket: (bucket_indices, representative_dyn_map, stitched_llir).
@@ -485,7 +338,9 @@ impl CompileOptions {
         log_channel_enabled(self.rolling_log, "ROLLING_LOG")
     }
 
-    fn search_log_enabled(&self) -> bool {
+    /// Whether search progress logging is on, honoring `SEARCH_LOG` /
+    /// `LUMINAL_LOG` overrides.
+    pub fn search_log_enabled(&self) -> bool {
         log_channel_enabled(self.search_log, "SEARCH_LOG")
     }
 }
@@ -533,83 +388,6 @@ fn validate_dim_buckets(dimension: Symbol, buckets: &[DimBucket]) {
     }
 }
 
-fn maybe_dump_selected_llir(label: &str, dyn_map: &DynMap, llir: &LLIRGraph) {
-    let Ok(dir) = std::env::var("LLIR_DUMP_DIR") else {
-        return;
-    };
-
-    if let Err(err) = std::fs::create_dir_all(&dir) {
-        eprintln!("failed to create LLIR_DUMP_DIR={dir}: {err}");
-        return;
-    }
-
-    let dims = dyn_map
-        .iter()
-        .sorted_by_key(|(dim, _)| **dim)
-        .map(|(dim, value)| format!("{dim}{value}"))
-        .join("_");
-    let stem = if dims.is_empty() {
-        format!("selected-llir-{label}")
-    } else {
-        format!("selected-llir-{label}-{dims}")
-    };
-    let dot_path = format!("{dir}/{stem}.dot");
-    let summary_path = format!("{dir}/{stem}.txt");
-
-    let dot = format!("{:?}", Dot::with_config(llir, &[Config::EdgeNoLabel]));
-    if let Err(err) = std::fs::write(&dot_path, dot) {
-        eprintln!("failed to write {dot_path}: {err}");
-    }
-
-    let mut op_counts = std::collections::BTreeMap::<String, usize>::new();
-    for node in llir.node_indices() {
-        *op_counts.entry(format!("{}", llir[node])).or_default() += 1;
-    }
-
-    let mut summary = String::new();
-    let _ = writeln!(
-        summary,
-        "selected LLIR {label}: {} nodes, {} edges",
-        llir.node_count(),
-        llir.edge_count()
-    );
-    let _ = writeln!(summary, "dyn_map: {dyn_map:?}");
-    let _ = writeln!(summary, "\nop counts:");
-    for (op, count) in op_counts {
-        let _ = writeln!(summary, "  {count:5} {op}");
-    }
-    let _ = writeln!(summary, "\nnodes:");
-    for node in llir.node_indices().sorted_by_key(|n| n.index()) {
-        let inputs = llir
-            .edges_directed(node, Direction::Incoming)
-            .sorted_by_key(|edge| edge.id())
-            .map(|edge| edge.source().index().to_string())
-            .join(", ");
-        let _ = writeln!(
-            summary,
-            "  n{} <- [{}] {}",
-            node.index(),
-            inputs,
-            llir[node]
-        );
-    }
-
-    if let Err(err) = std::fs::write(&summary_path, summary) {
-        eprintln!("failed to write {summary_path}: {err}");
-    } else {
-        println!("   LLIR dump {summary_path}");
-    }
-}
-
-fn panic_initial_filter_limit(filter_fails: usize, last_rejection: Option<&str>) -> ! {
-    if let Some(last_rejection) = last_rejection {
-        panic!(
-            "Failed to find a viable initial genome after {filter_fails} runtime filter failures; last rejection: {last_rejection}"
-        );
-    }
-    panic!("Failed to find a viable initial genome after {filter_fails} runtime filter failures");
-}
-
 /// A Luminal compute graph.
 ///
 /// All computation is represented as a directed acyclic graph.
@@ -619,16 +397,11 @@ pub struct Graph {
     pub dyn_map: DynMap,
     /// Edge weights: (Input index, Output index, Input shape)
     pub graph: HLIRGraph,
-    /// E-Graph search spaces. Bucketed compilation stores one egraph per
-    /// bucket combination; unbucketed compilation stores one egraph.
-    egraphs: Vec<SerializedEGraph>,
-    egraph_contexts: Vec<SearchSpaceContext>,
-    /// Available ops
-    pub ops: Option<Vec<Arc<Box<dyn EgglogOp>>>>,
+    /// The saturated search space built by [`Graph::build_search_space`]:
+    /// one e-graph per bucket combination, handed to the runtime to search.
+    search_space: Option<SearchSpace>,
     /// Custom ops
     pub custom_ops: Vec<Box<dyn CustomOp>>,
-    /// Bucket definitions used by the currently built search space.
-    search_space_dim_buckets: FxHashMap<Symbol, Vec<DimBucket>>,
     /// Optional graph-wide interval assumptions for dynamic dimensions.
     pub dim_intervals: DynDimIntervals,
     /// Metadata for Input nodes: NodeIndex -> (label, dtype).
@@ -1786,83 +1559,101 @@ impl Graph {
     }
 
     #[tracing::instrument(skip_all)]
-    pub fn build_search_space<Rt: Runtime + 'static>(&mut self, options: CompileOptions) {
-        self.run_auto_loop_rolling_prepass(&options);
+    pub fn build_search_space<Rt: Runtime>(&mut self, options: CompileOptions) {
         let mut ops = Rt::Ops::into_vec();
         ops.extend(<crate::hlir::HLIROps as IntoEgglogOp>::into_vec());
-        let cleanup_hlir = TypeId::of::<Rt>() != TypeId::of::<ReferenceRuntime>();
+        self.build_search_space_with_ops::<Rt>(ops, options);
+    }
+
+    #[tracing::instrument(skip_all)]
+    pub fn build_search_space_exclude_ops<Rt: Runtime, Ex: IntoEgglogOp>(
+        &mut self,
+        options: CompileOptions,
+    ) {
+        let exclude_ops = Ex::into_vec()
+            .into_iter()
+            .map(|e| e.sort().name)
+            .collect::<FxHashSet<_>>();
+        let mut ops = Rt::Ops::into_vec();
+        ops.retain(|o| !exclude_ops.contains(&o.sort().name));
+        ops.extend(<crate::hlir::HLIROps as IntoEgglogOp>::into_vec());
+        self.build_search_space_with_ops::<Rt>(ops, options);
+    }
+
+    /// Roll loops, then saturate one e-graph per bucket combination into a
+    /// [`SearchSpace`] the runtime searches in [`Runtime::compile`].
+    fn build_search_space_with_ops<Rt: Runtime>(
+        &mut self,
+        ops: Vec<Arc<Box<dyn EgglogOp>>>,
+        options: CompileOptions,
+    ) {
+        self.run_auto_loop_rolling_prepass(&options);
         let dim_buckets = options.dim_buckets.clone();
         let late_pass_dyn_map = self.late_pass_dyn_map(&dim_buckets);
         let late_passes = Rt::late_egglog_passes(&ops, &options, &late_pass_dyn_map);
         let extra_egglog = Rt::extra_egglog();
 
         let (program, root) = hlir_to_egglog(self);
-        let contexts = self.search_space_contexts(&dim_buckets);
-        self.egraphs = contexts
-            .iter()
-            .map(|context| {
+        let buckets = bucket_index_combinations(&dim_buckets)
+            .into_iter()
+            .map(|bucket_indices| {
+                let intervals = self.bucket_intervals(&dim_buckets, &bucket_indices);
                 let (contextual_program, use_interval_analysis) =
-                    self.egglog_program_with_interval_facts(&program, &context.intervals);
-                run_egglog_with_late_passes_interval_analysis_and_log(
+                    self.egglog_program_with_interval_facts(&program, &intervals);
+                let egraph = run_egglog_with_late_passes_interval_analysis_and_log(
                     &contextual_program,
                     &root,
                     &ops,
-                    cleanup_hlir,
+                    Rt::CLEANUP_HLIR,
                     &late_passes,
                     &extra_egglog,
                     use_interval_analysis,
                     options.egglog_log_enabled(),
                 )
-                .unwrap()
-            })
-            .collect();
-        self.egraph_contexts = contexts;
-        self.ops = Some(ops);
-        self.search_space_dim_buckets = dim_buckets;
-    }
-
-    fn search_space_contexts(
-        &self,
-        dim_buckets: &FxHashMap<Symbol, Vec<DimBucket>>,
-    ) -> Vec<SearchSpaceContext> {
-        if dim_buckets.is_empty() {
-            return vec![SearchSpaceContext {
-                bucket_indices: FxHashMap::default(),
-                representative_dyn_map: self.dyn_map.clone(),
-                intervals: self.dim_intervals.clone(),
-            }];
-        }
-
-        self.bucket_combinations(dim_buckets)
-            .into_iter()
-            .map(|(bucket_indices, representative_dyn_map)| {
-                let mut intervals = self.dim_intervals.clone();
-                for (&dim, &idx) in &bucket_indices {
-                    let bucket = &dim_buckets[&dim][idx];
-                    let min = i64::try_from(bucket.min)
-                        .expect("DimBucket min must fit into i64 for interval analysis");
-                    let max = i64::try_from(bucket.max)
-                        .expect("DimBucket max must fit into i64 for interval analysis");
-                    let bucket_interval = DimInterval::new(min, max);
-                    intervals
-                        .entry(dim)
-                        .and_modify(|existing| {
-                            existing.min = existing.min.max(bucket_interval.min);
-                            existing.max = existing.max.min(bucket_interval.max);
-                            assert!(
-                                existing.min <= existing.max,
-                                "Bucket interval for dim '{dim}' does not overlap graph interval"
-                            );
-                        })
-                        .or_insert(bucket_interval);
-                }
-                SearchSpaceContext {
+                .unwrap();
+                BucketSearchSpace {
+                    egraph,
                     bucket_indices,
-                    representative_dyn_map,
                     intervals,
                 }
             })
-            .collect()
+            .collect();
+        let custom_ops = self.custom_ops.iter().map(|op| op.to_llir_op()).collect();
+        self.search_space = Some(SearchSpace {
+            buckets,
+            ops,
+            custom_ops,
+            dim_buckets,
+        });
+    }
+
+    /// Graph-wide interval assumptions narrowed to one bucket combination.
+    fn bucket_intervals(
+        &self,
+        dim_buckets: &FxHashMap<Symbol, Vec<DimBucket>>,
+        bucket_indices: &DynMap,
+    ) -> DynDimIntervals {
+        let mut intervals = self.dim_intervals.clone();
+        for (&dim, &idx) in bucket_indices {
+            let bucket = &dim_buckets[&dim][idx];
+            let min = i64::try_from(bucket.min)
+                .expect("DimBucket min must fit into i64 for interval analysis");
+            let max = i64::try_from(bucket.max)
+                .expect("DimBucket max must fit into i64 for interval analysis");
+            let bucket_interval = DimInterval::new(min, max);
+            intervals
+                .entry(dim)
+                .and_modify(|existing| {
+                    existing.min = existing.min.max(bucket_interval.min);
+                    existing.max = existing.max.min(bucket_interval.max);
+                    assert!(
+                        existing.min <= existing.max,
+                        "Bucket interval for dim '{dim}' does not overlap graph interval"
+                    );
+                })
+                .or_insert(bucket_interval);
+        }
+        intervals
     }
 
     fn egglog_program_with_interval_facts(
@@ -1878,58 +1669,33 @@ impl Graph {
         }
     }
 
-    #[tracing::instrument(skip_all)]
-    pub fn build_search_space_exclude_ops<Rt: Runtime + 'static, Ex: IntoEgglogOp>(
-        &mut self,
-        options: CompileOptions,
-    ) {
-        self.run_auto_loop_rolling_prepass(&options);
-        let exclude_ops = Ex::into_vec()
-            .into_iter()
-            .map(|e| e.sort().name)
-            .collect::<FxHashSet<_>>();
-        let mut ops = Rt::Ops::into_vec();
-        ops.retain(|o| !exclude_ops.contains(&o.sort().name));
-        ops.extend(<crate::hlir::HLIROps as IntoEgglogOp>::into_vec());
-        let cleanup_hlir = TypeId::of::<Rt>() != TypeId::of::<ReferenceRuntime>();
-        let dim_buckets = options.dim_buckets.clone();
-        let late_pass_dyn_map = self.late_pass_dyn_map(&dim_buckets);
-        let late_passes = Rt::late_egglog_passes(&ops, &options, &late_pass_dyn_map);
-        let extra_egglog = Rt::extra_egglog();
-
-        let (program, root) = hlir_to_egglog(self);
-        let contexts = self.search_space_contexts(&dim_buckets);
-        self.egraphs = contexts
-            .iter()
-            .map(|context| {
-                let (contextual_program, use_interval_analysis) =
-                    self.egglog_program_with_interval_facts(&program, &context.intervals);
-                run_egglog_with_late_passes_interval_analysis_and_log(
-                    &contextual_program,
-                    &root,
-                    &ops,
-                    cleanup_hlir,
-                    &late_passes,
-                    &extra_egglog,
-                    use_interval_analysis,
-                    options.egglog_log_enabled(),
-                )
-                .unwrap()
-            })
-            .collect();
-        self.egraph_contexts = contexts;
-        self.ops = Some(ops);
-        self.search_space_dim_buckets = dim_buckets;
+    /// Dyn map handed to backend late passes: bucket maxima override the
+    /// graph's values so passes plan for the largest shape.
+    fn late_pass_dyn_map(&self, dim_buckets: &FxHashMap<Symbol, Vec<DimBucket>>) -> DynMap {
+        let mut dyn_map = self.dyn_map.clone();
+        for (&dim, buckets) in dim_buckets {
+            if let Some(max) = buckets.iter().map(|bucket| bucket.max).max() {
+                dyn_map.insert(dim, max);
+            }
+        }
+        dyn_map
+    }
+    /// The built search space, if [`Graph::build_search_space`] has run.
+    pub fn search_space(&self) -> Option<&SearchSpace> {
+        self.search_space.as_ref()
     }
 
     /// Get a reference to the first e-graph search space (if built)
     pub fn egraph(&self) -> Option<&SerializedEGraph> {
-        self.egraphs.first()
+        self.search_space
+            .as_ref()
+            .and_then(|space| space.buckets.first())
+            .map(|bucket| &bucket.egraph)
     }
 
     /// Get a reference to the available ops (if search space is built)
     pub fn egglog_ops(&self) -> Option<&Vec<Arc<Box<dyn EgglogOp>>>> {
-        self.ops.as_ref()
+        self.search_space.as_ref().map(|space| &space.ops)
     }
 
     /// Build the search space and search it with one shared set of options.
@@ -1938,13 +1704,13 @@ impl Graph {
     /// weights have already been loaded. Use `build_search_space` and `search`
     /// directly when the two phases need to be separated.
     #[tracing::instrument(skip_all)]
-    pub fn compile<R: Runtime + 'static>(&mut self, runtime: R, options: CompileOptions) -> R {
+    pub fn compile<R: Runtime>(&mut self, runtime: R, options: CompileOptions) -> R {
         let mut rng = rand::rng();
         self.compile_with_rng(runtime, options, &mut rng)
     }
 
     #[tracing::instrument(skip_all)]
-    pub fn compile_with_rng<R: Runtime + 'static, G: rand::Rng>(
+    pub fn compile_with_rng<R: Runtime, G: rand::Rng>(
         &mut self,
         runtime: R,
         options: CompileOptions,
@@ -1962,13 +1728,15 @@ impl Graph {
     }
 
     #[tracing::instrument(skip_all)]
-    pub fn search<R: Runtime + 'static>(&mut self, runtime: R, options: CompileOptions) -> R {
+    pub fn search<R: Runtime>(&mut self, runtime: R, options: CompileOptions) -> R {
         let mut rng = rand::rng();
         self.search_with_rng(runtime, options, &mut rng)
     }
 
+    /// Hand the built search space to the runtime, which searches it by
+    /// whatever strategy it implements and loads the programs it selects.
     #[tracing::instrument(skip_all)]
-    pub fn search_with_rng<R: Runtime + 'static, G: rand::Rng>(
+    pub fn search_with_rng<R: Runtime, G: rand::Rng>(
         &mut self,
         mut runtime: R,
         options: CompileOptions,
@@ -1977,1123 +1745,18 @@ impl Graph {
         for (&dim, &value) in &options.search_dims {
             self.set_dim(dim, value);
         }
-
+        let space = self
+            .search_space
+            .as_ref()
+            .expect("build_search_space must run before search");
         assert!(
-            options.dim_buckets.is_empty() || options.dim_buckets == self.search_space_dim_buckets,
+            options.dim_buckets.is_empty() || options.dim_buckets == space.dim_buckets,
             "dim buckets must be configured in CompileOptions before build_search_space; search cannot change buckets after build",
         );
-
-        let search_started_at = std::time::Instant::now();
-        let search_log = options.search_log_enabled();
-        if self.search_space_dim_buckets.is_empty() {
-            // No buckets: existing single-search path
-            let ranked = self.search_single(
-                &mut runtime,
-                &options,
-                rng,
-                &self.dyn_map.clone(),
-                None,
-                None,
-                0,
-                search_started_at,
-            );
-            let mut candidates = LazyFinalists::new(ranked);
-            if !self.ensure_finalist(
-                &mut runtime,
-                &mut candidates,
-                0,
-                &options,
-                &self.dyn_map,
-                None,
-                0,
-                search_started_at,
-            ) {
-                panic!(
-                    "Failed to find a viable final graph: {}",
-                    Self::no_finalist_message(&candidates)
-                );
-            }
-            let finalist = candidates.finalists.remove(0);
-            Self::dump_selected_finalist(&finalist, &self.dyn_map, None);
-
-            runtime.clear_intermediate_buffers();
-            runtime.load_llir(&finalist.llir);
-            runtime
-        } else {
-            // Bucketed search: retain ranked genomes for every bucket, then
-            // choose a collectively viable finalist set below.
-            let bucket_contexts = self.search_space_contexts(&self.search_space_dim_buckets);
-            let n_combos = bucket_contexts.len();
-            let mut bucket_searches = Vec::with_capacity(n_combos);
-            assert!(
-                self.egraphs.len() == n_combos,
-                "dim buckets must be configured before build_search_space; search space has {} egraphs but current bucket configuration has {n_combos} combinations",
-                self.egraphs.len(),
-            );
-
-            for (combo_idx, context) in bucket_contexts.into_iter().enumerate() {
-                let bucket_label = self
-                    .format_bucket_label(&self.search_space_dim_buckets, &context.bucket_indices);
-                if search_log {
-                    println!(
-                        "   {:>6}  Group {}/{}: {}",
-                        "Search".cyan().bold(),
-                        combo_idx + 1,
-                        n_combos,
-                        bucket_label,
-                    );
-                }
-
-                let profile_context = SearchProfileBucketContext {
-                    dim_buckets: self.search_space_dim_buckets.clone(),
-                    bucket_indices: context.bucket_indices.clone(),
-                    representative_dyn_map: context.representative_dyn_map.clone(),
-                };
-                let ranked = self.search_single(
-                    &mut runtime,
-                    &options,
-                    rng,
-                    &context.representative_dyn_map,
-                    Some((combo_idx, n_combos)),
-                    Some(profile_context.clone()),
-                    combo_idx,
-                    search_started_at,
-                );
-                bucket_searches.push(BucketFinalistSearch {
-                    context: profile_context,
-                    egraph_index: combo_idx,
-                    candidates: LazyFinalists::new(ranked),
-                });
-            }
-
-            // Materialize only the fastest individually viable graph for each
-            // bucket to seed a best-first walk over the Cartesian finalist
-            // lattice. Slower full LLIRs are extracted only when an aggregate
-            // rejection makes their coordinate reachable.
-            for (bucket_idx, bucket) in bucket_searches.iter_mut().enumerate() {
-                if !self.ensure_finalist(
-                    &mut runtime,
-                    &mut bucket.candidates,
-                    0,
-                    &options,
-                    &bucket.context.representative_dyn_map,
-                    Some(&bucket.context),
-                    bucket.egraph_index,
-                    search_started_at,
-                ) {
-                    let label = self.format_bucket_label(
-                        &self.search_space_dim_buckets,
-                        &bucket.context.bucket_indices,
-                    );
-                    panic!(
-                        "Failed to find a viable final graph for bucket {bucket_idx} ({label}): {}",
-                        Self::no_finalist_message(&bucket.candidates)
-                    );
-                }
-            }
-
-            let initial_indices = vec![0usize; bucket_searches.len()];
-            let initial_metrics = bucket_searches
-                .iter()
-                .map(|bucket| bucket.candidates.finalists[0].metric.clone())
-                .collect_vec();
-            let mut frontier = vec![(
-                R::aggregate_profile_metrics(&initial_metrics),
-                initial_indices.clone(),
-            )];
-            let mut visited = FxHashSet::default();
-            visited.insert(initial_indices);
-            let mut selected_indices = None;
-            let mut aggregate_attempts = 0usize;
-            let mut aggregate_rejections = 0usize;
-            let mut last_aggregate_rejection = None;
-            let mut aggregate_stopped_reason = None;
-
-            while !frontier.is_empty() {
-                if aggregate_attempts > 0
-                    && search_started_at.elapsed() >= options.search_time_limit
-                {
-                    aggregate_stopped_reason = Some(
-                        "search time limit expired during aggregate bucket finalization"
-                            .to_string(),
-                    );
-                    break;
-                }
-                let best_pos = (1..frontier.len()).fold(0, |best, candidate| {
-                    if frontier[candidate]
-                        .0
-                        .partial_cmp(&frontier[best].0)
-                        .is_some_and(|ordering| ordering == std::cmp::Ordering::Less)
-                    {
-                        candidate
-                    } else {
-                        best
-                    }
-                });
-                let (_, indices) = frontier.swap_remove(best_pos);
-                aggregate_attempts += 1;
-
-                let bucket_refs = bucket_searches
-                    .iter()
-                    .zip(&indices)
-                    .map(|(bucket, &candidate_idx)| BucketLLIRRef {
-                        bucket_indices: &bucket.context.bucket_indices,
-                        representative_dyn_map: &bucket.context.representative_dyn_map,
-                        llir: &bucket.candidates.finalists[candidate_idx].llir,
-                    })
-                    .collect_vec();
-                runtime.clear_intermediate_buffers();
-                let aggregate_started_at = std::time::Instant::now();
-                let aggregate_result =
-                    std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-                        runtime.filter_llir_bucket_set(
-                            &self.search_space_dim_buckets,
-                            &bucket_refs,
-                            &options,
-                        )
-                    }))
-                    .unwrap_or_else(|_| {
-                        CandidateFilterResult::reject_with_display(
-                            "aggregate bucket candidate filter panicked",
-                        )
-                    });
-                let aggregate_timed_out = options
-                    .candidate_timeout
-                    .is_some_and(|timeout| aggregate_started_at.elapsed() >= timeout);
-                if aggregate_result.accepted && !aggregate_timed_out {
-                    if search_log && aggregate_rejections > 0 {
-                        println!(
-                            "   {:>6}  aggregate fallback: selected per-bucket finalist ranks {:?} after {} rejection(s)",
-                            "Search".yellow().bold(),
-                            indices,
-                            aggregate_rejections,
-                        );
-                    }
-                    selected_indices = Some(indices);
-                    break;
-                }
-                aggregate_rejections += 1;
-                crate::mask_events::AGGREGATE_REJECT.record();
-                last_aggregate_rejection = if aggregate_timed_out {
-                    Some(format!(
-                        "candidate timeout expired while filtering aggregate bucket set {aggregate_attempts}"
-                    ))
-                } else {
-                    aggregate_result.display
-                };
-                if search_log {
-                    println!(
-                        "   {:>6}  aggregate reject per-bucket finalist ranks {:?}: {}",
-                        "Search".yellow().bold(),
-                        indices,
-                        last_aggregate_rejection.as_deref().unwrap_or("(no reason)"),
-                    );
-                }
-
-                // Any one-coordinate successor is the next possible slower
-                // combination. A visited set prevents duplicate lattice paths.
-                for bucket_idx in 0..bucket_searches.len() {
-                    let mut successor = indices.clone();
-                    successor[bucket_idx] += 1;
-                    if !visited.insert(successor.clone()) {
-                        continue;
-                    }
-                    let bucket = &mut bucket_searches[bucket_idx];
-                    if !self.ensure_finalist(
-                        &mut runtime,
-                        &mut bucket.candidates,
-                        successor[bucket_idx],
-                        &options,
-                        &bucket.context.representative_dyn_map,
-                        Some(&bucket.context),
-                        bucket.egraph_index,
-                        search_started_at,
-                    ) {
-                        if bucket.candidates.stopped_reason.is_some()
-                            || bucket
-                                .candidates
-                                .last_rejection
-                                .as_deref()
-                                .is_some_and(|reason| reason.contains("timeout"))
-                        {
-                            let label = self.format_bucket_label(
-                                &self.search_space_dim_buckets,
-                                &bucket.context.bucket_indices,
-                            );
-                            aggregate_stopped_reason.get_or_insert_with(|| {
-                                format!(
-                                    "bucket {bucket_idx} ({label}) fallback stopped: {}",
-                                    Self::no_finalist_message(&bucket.candidates)
-                                )
-                            });
-                        }
-                        continue;
-                    }
-                    let metrics = bucket_searches
-                        .iter()
-                        .zip(&successor)
-                        .map(|(bucket, &candidate_idx)| {
-                            bucket.candidates.finalists[candidate_idx].metric.clone()
-                        })
-                        .collect_vec();
-                    frontier.push((R::aggregate_profile_metrics(&metrics), successor));
-                }
-            }
-
-            let Some(selected_indices) = selected_indices else {
-                let reason = aggregate_stopped_reason
-                    .or(last_aggregate_rejection)
-                    .unwrap_or_else(|| "no aggregate candidate combinations remain".to_string());
-                panic!(
-                    "Failed to find a viable aggregate bucket set after {aggregate_rejections} rejections: {reason}"
-                );
-            };
-
-            let mut bucket_llirs = Vec::with_capacity(n_combos);
-            for (bucket_idx, (mut bucket, candidate_idx)) in bucket_searches
-                .into_iter()
-                .zip(selected_indices)
-                .enumerate()
-            {
-                let finalist = bucket.candidates.finalists.swap_remove(candidate_idx);
-                Self::dump_selected_finalist(
-                    &finalist,
-                    &bucket.context.representative_dyn_map,
-                    Some((bucket_idx, n_combos)),
-                );
-                bucket_llirs.push((
-                    bucket.context.bucket_indices,
-                    bucket.context.representative_dyn_map,
-                    finalist.llir,
-                ));
-            }
-
-            runtime.clear_intermediate_buffers();
-            runtime.load_llir_buckets(&self.search_space_dim_buckets, &bucket_llirs);
-            runtime
-        }
-    }
-
-    /// Compute cartesian product of all bucket combinations.
-    /// Returns Vec of (bucket_indices, representative_dyn_map).
-    fn bucket_combinations(
-        &self,
-        dim_buckets: &FxHashMap<Symbol, Vec<DimBucket>>,
-    ) -> Vec<(DynMap, DynMap)> {
-        let mut dims: Vec<(Symbol, &Vec<DimBucket>)> =
-            dim_buckets.iter().map(|(c, b)| (*c, b)).collect();
-        dims.sort_by_key(|(c, _)| *c);
-
-        let mut combos: Vec<(DynMap, DynMap)> = vec![(FxHashMap::default(), self.dyn_map.clone())];
-
-        for (dim, buckets) in &dims {
-            let mut new_combos = Vec::new();
-            for (existing_indices, existing_dyn_map) in &combos {
-                for (bucket_idx, bucket) in buckets.iter().enumerate() {
-                    let mut indices = existing_indices.clone();
-                    indices.insert(*dim, bucket_idx);
-                    let mut dyn_map = existing_dyn_map.clone();
-                    dyn_map.insert(*dim, bucket.representative_value());
-                    new_combos.push((indices, dyn_map));
-                }
-            }
-            combos = new_combos;
-        }
-
-        combos
-    }
-
-    fn late_pass_dyn_map(&self, dim_buckets: &FxHashMap<Symbol, Vec<DimBucket>>) -> DynMap {
-        let mut dyn_map = self.dyn_map.clone();
-        for (&dim, buckets) in dim_buckets {
-            if let Some(max) = buckets.iter().map(|bucket| bucket.max).max() {
-                dyn_map.insert(dim, max);
-            }
-        }
-        dyn_map
-    }
-
-    /// Format a human-readable label for a bucket combination.
-    fn format_bucket_label(
-        &self,
-        dim_buckets: &FxHashMap<Symbol, Vec<DimBucket>>,
-        bucket_indices: &DynMap,
-    ) -> String {
-        let mut parts: Vec<String> = Vec::new();
-        let mut dims: Vec<_> = bucket_indices.iter().collect();
-        dims.sort_by_key(|(c, _)| **c);
-        for (dim, &idx) in dims {
-            let bucket = &dim_buckets[dim][idx];
-            if bucket.min == bucket.max {
-                parts.push(format!("{}={}", dim, bucket.min));
-            } else {
-                parts.push(format!(
-                    "{}=[{},{}]@{}",
-                    dim,
-                    bucket.min,
-                    bucket.max,
-                    bucket.representative_value()
-                ));
-            }
-        }
-        parts.join(", ")
-    }
-
-    /// Run the genetic search and return every successfully profiled genome in
-    /// metric order. Final extraction and hard filtering are deliberately
-    /// deferred so bucketed compilation can choose a viable retained set.
-    /// `bucket_progress`: if `Some((current_bucket_idx, total_buckets))` adds a
-    /// second "Bucket" progress bar.
-    #[allow(clippy::too_many_arguments)]
-    fn search_single<R: Runtime + 'static, G: rand::Rng>(
-        &self,
-        runtime: &mut R,
-        options: &CompileOptions,
-        rng: &mut G,
-        dyn_map: &DynMap,
-        bucket_progress: Option<(usize, usize)>,
-        bucket_profile_context: Option<SearchProfileBucketContext>,
-        egraph_index: usize,
-        search_started_at: std::time::Instant,
-    ) -> Vec<(R::ProfileMetric, IndexedChoiceSet)> {
-        let mut profile_dyn_map = dyn_map.clone();
-        for (&dim, &value) in &options.profile_dims {
-            profile_dyn_map.insert(dim, value);
-        }
-        let limit = options.limit;
-        let ops = self.ops.as_ref().unwrap();
-        let egraph = &self.egraphs[egraph_index];
-        let search_limit = count_choice_sets_up_to(egraph, limit);
-        let start = std::time::Instant::now();
-        let search_log = options.search_log_enabled();
-
-        // Bar layout: one Search bar, plus an optional Bucket bar.
-        let n_bar_lines = 1 + if bucket_progress.is_some() { 1 } else { 0 };
-
-        fn make_bar(searched: usize, total: usize) -> String {
-            let bar_width = 24;
-            let head = ((searched as f32 / total as f32) * bar_width as f32)
-                .clamp(0.0, bar_width as f32)
-                .floor() as usize;
-            if head == 0 {
-                format!("[>{}]", " ".repeat(bar_width - 1))
-            } else if head >= bar_width {
-                format!("[{}>]", "=".repeat(bar_width))
-            } else {
-                format!(
-                    "[{}>{}]",
-                    "=".repeat(head),
-                    " ".repeat(bar_width - head - 1)
-                )
-            }
-        }
-
-        let render_bars =
-            |n_graphs: usize, limit: usize, bucket_progress: Option<(usize, usize)>| {
-                print!(
-                    "\x1b[2K  {:>6}  {} {n_graphs}/{limit}",
-                    "Search".cyan().bold(),
-                    make_bar(n_graphs, limit),
-                );
-                if let Some((bucket_idx, n_buckets)) = bucket_progress {
-                    print!(
-                        "\n\x1b[2K  {:>6}  {} {}/{n_buckets}",
-                        "Bucket".cyan().bold(),
-                        make_bar(bucket_idx, n_buckets),
-                        bucket_idx,
-                    );
-                }
-            };
-
-        let mut prev_selected: FxHashSet<u64> = FxHashSet::default();
-        let mut explored_llir_hashes: FxHashSet<LlirFingerprint> = FxHashSet::default();
-        let mut llir_extractor = LlirExtractor::new(egraph, ops);
-        runtime.clear_intermediate_buffers();
-        let search_time_limit_reached = || search_started_at.elapsed() >= options.search_time_limit;
-        // `profile_start.elapsed()` wraps the backend's post-acceptance profile
-        // phase. Hard filtering/preparation is governed by the overall search
-        // time limit; `execution_timeout` separately bounds profiling runs.
-        let candidate_timed_out = |elapsed: std::time::Duration| {
-            options
-                .candidate_timeout
-                .is_some_and(|timeout| elapsed >= timeout)
-        };
-
-        // Find a viable initial genome. Runtime-filtered candidates are dry
-        // failures, not searched graphs: they are never profiled and do not
-        // count toward the graph search limit.
-        let (initial_genome, mut best_metric, display, mut n_graphs) = {
-            let mut invalid_attempts = 0usize;
-            let mut filter_fails = 0usize;
-            let mut last_filter_rejection: Option<String> = None;
-            // Breakdown of why profiled candidates were invalid, for the give-up message.
-            let (mut n_timed_out, mut n_invalid_profile, mut n_panicked) = (0, 0, 0);
-            let max_invalid_attempts = 100usize;
-            let max_filter_fails = options
-                .limit
-                .max(1)
-                .saturating_mul(options.generation_size.max(1))
-                .saturating_mul(100)
-                .max(10_000);
-
-            loop {
-                let mut generation =
-                    llir_extractor.random_indexed_generation(1, &mut prev_selected, rng);
-                let Some(genome) = generation.pop() else {
-                    panic_initial_filter_limit(filter_fails, last_filter_rejection.as_deref());
-                };
-
-                let graph_result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-                    let packed =
-                        llir_extractor.extract_indexed_packed(&genome, &self.custom_ops, None);
-                    if !explored_llir_hashes.insert(packed.fingerprint()) {
-                        return None;
-                    }
-                    // Profile the deployment graph itself: fully unrolled.
-                    // Every scaled-down proxy (collapsed bodies, trip-count
-                    // differencing) leaked family-dependent costs and
-                    // inverted rankings; measuring the real graph is slower
-                    // per candidate but cannot misorder families.
-                    Some(unroll_packed_llir(packed))
-                }));
-                let graph = match graph_result {
-                    Ok(Some(graph)) => graph,
-                    Ok(None) => continue,
-                    Err(_) => {
-                        invalid_attempts += 1;
-                        if invalid_attempts > max_invalid_attempts {
-                            panic!(
-                                "Failed to find a viable initial genome after {max_invalid_attempts} invalid attempts"
-                            );
-                        }
-                        continue;
-                    }
-                };
-
-                // A candidate whose LLIR fails to compile (e.g. an egglog
-                // rule that mis-fires and produces an inconsistent kernel op)
-                // must be rejected like any other, not abort the search.
-                let filter_result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-                    self.prepare_profile_candidate(
-                        runtime,
-                        &graph,
-                        &profile_dyn_map,
-                        options,
-                        bucket_profile_context.as_ref(),
-                    )
-                }))
-                .unwrap_or_else(|_| {
-                    CandidateFilterResult::reject_with_display("candidate compile panicked")
-                });
-                if !filter_result.accepted {
-                    filter_fails += 1;
-                    last_filter_rejection = filter_result.display;
-                    // Rejections are otherwise silent until the 10k-fail
-                    // panic; surface them early — a structural rejection
-                    // (e.g. every candidate over the memory cap) loops
-                    // here for hours looking like a hang.
-                    if filter_fails <= 5 || filter_fails % 100 == 0 {
-                        eprintln!(
-                            "   Search  initial-genome filter reject #{filter_fails}: {}",
-                            last_filter_rejection.as_deref().unwrap_or("(no reason)")
-                        );
-                    }
-                    if filter_fails >= max_filter_fails {
-                        panic_initial_filter_limit(filter_fails, last_filter_rejection.as_deref());
-                    }
-                    continue;
-                }
-                let filter_display = filter_result.display;
-
-                let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-                    let profile_start = std::time::Instant::now();
-                    let (rep_metric, rep_display) = runtime.profile_prepared_candidate(
-                        &graph,
-                        &profile_dyn_map,
-                        options.trials,
-                        options.execution_timeout,
-                        // No best exists yet — the initial genome establishes
-                        // the early-stop baseline.
-                        None,
-                        bucket_profile_context.as_ref().map(|bucket_context| {
-                            ProfileBucketContext {
-                                dim_buckets: &bucket_context.dim_buckets,
-                                bucket_indices: &bucket_context.bucket_indices,
-                                representative_dyn_map: &bucket_context.representative_dyn_map,
-                            }
-                        }),
-                    );
-                    let timed_out = candidate_timed_out(profile_start.elapsed());
-                    let invalid_profile = rep_display.starts_with("invalid ");
-                    if !timed_out && !invalid_profile {
-                        log_best_llir(&graph, &format!("candidate=0 {rep_display}"));
-                    }
-                    (
-                        rep_metric,
-                        append_filter_display(rep_display, filter_display.as_deref()),
-                        timed_out,
-                        invalid_profile,
-                    )
-                }));
-
-                match result {
-                    Ok((metric, disp, false, false)) => {
-                        break (genome, R::aggregate_profile_metrics(&[metric]), disp, 1);
-                    }
-                    Ok((_, _, timed_out, invalid_profile)) => {
-                        if timed_out {
-                            n_timed_out += 1;
-                        } else if invalid_profile {
-                            n_invalid_profile += 1;
-                        }
-                        invalid_attempts += 1;
-                    }
-                    Err(_) => {
-                        n_panicked += 1;
-                        invalid_attempts += 1;
-                    }
-                }
-                if invalid_attempts > max_invalid_attempts {
-                    panic!(
-                        "Failed to find a viable initial genome after {max_invalid_attempts} invalid attempts \
-                         (candidate_timed_out={n_timed_out} invalid_profile={n_invalid_profile} panicked={n_panicked})"
-                    );
-                }
-            }
-        };
-
-        // Print initial result and progress
-        if search_log {
-            let msg = format!("   {:>6} {}", "Start".cyan().bold(), display);
-            println!("{msg}");
-            render_bars(n_graphs, search_limit, bucket_progress);
-            std::io::stdout().flush().unwrap();
-        }
-
-        // Retain every successfully profiled genome in metric order. Profiling
-        // uses a collapsed loop body, so the fastest collapsed candidate is not
-        // necessarily viable after the final loop unroll. Final selection below
-        // re-extracts and hard-filters these candidates fastest-first.
-        let mut ranked_candidates = vec![(best_metric.clone(), initial_genome.clone())];
-
-        // Track top-N parents for offspring generation
-        let mut parents: Vec<(R::ProfileMetric, IndexedChoiceSet)> =
-            vec![(best_metric.clone(), initial_genome)];
-        let mut resample_generation = false;
-        let mut stagnant_generations = 0usize;
-        let mut slower_since_faster = 0usize;
-        let mut slower_line_visible = false;
-
-        while n_graphs < search_limit {
-            if search_time_limit_reached() {
-                break;
-            }
-
-            // Generate offspring from all parents, dividing budget evenly
-            let budget = (search_limit - n_graphs).min(options.generation_size);
-            let all_offspring = if resample_generation {
-                llir_extractor.random_indexed_generation(budget, &mut prev_selected, rng)
-            } else {
-                let per_parent = budget.div_ceil(parents.len());
-                let mut offspring = Vec::new();
-                for (_, parent_genome) in &parents {
-                    let remaining = budget.saturating_sub(offspring.len());
-                    if remaining == 0 {
-                        break;
-                    }
-                    // Stagnation kick: escaping a family basin needs
-                    // multi-gene jumps, so mutation counts escalate with
-                    // consecutive stagnant generations (capped 16x).
-                    let kick = if options.restart_stagnation > 0
-                        && stagnant_generations >= options.restart_stagnation
-                    {
-                        (1 + stagnant_generations - options.restart_stagnation).min(16)
-                    } else {
-                        1
-                    };
-                    offspring.extend(llir_extractor.extract_reachable_indexed_generation(
-                        parent_genome,
-                        per_parent.min(remaining),
-                        options.mutations * kick,
-                        &mut prev_selected,
-                        rng,
-                    ));
-                }
-                offspring
-            };
-            if all_offspring.is_empty() {
-                break;
-            }
-
-            let mut generation_found_non_timeout = false;
-            let mut generation_found_new_best = false;
-
-            for genome in all_offspring {
-                if search_time_limit_reached() {
-                    break;
-                }
-                let graph_result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-                    let packed =
-                        llir_extractor.extract_indexed_packed(&genome, &self.custom_ops, None);
-                    if !explored_llir_hashes.insert(packed.fingerprint()) {
-                        return None;
-                    }
-                    let pre_collapse = std::env::var_os("LLIR_DUMP_DIR")
-                        .is_some()
-                        .then(|| packed.to_stable());
-                    // Profile fully unrolled — see initial-genome path.
-                    let llir_graph = unroll_packed_llir(packed);
-                    Some((pre_collapse, llir_graph))
-                }));
-                if let Err(payload) = &graph_result {
-                    crate::mask_events::CANDIDATE_PANIC
-                        .record_with(|| crate::mask_events::panic_payload(payload.as_ref()));
-                }
-                let (pre_collapse, llir_graph) = match graph_result {
-                    Ok(Some(graphs)) => graphs,
-                    Ok(None) => continue,
-                    Err(_) => {
-                        if search_log {
-                            for _ in 1..n_bar_lines {
-                                print!("\x1b[1A");
-                            }
-                            print!("\r\x1b[2K");
-                            render_bars(n_graphs, search_limit, bucket_progress);
-                            std::io::stdout().flush().unwrap();
-                        }
-                        continue;
-                    }
-                };
-
-                let filter_result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-                    self.prepare_profile_candidate(
-                        runtime,
-                        &llir_graph,
-                        &profile_dyn_map,
-                        options,
-                        bucket_profile_context.as_ref(),
-                    )
-                }))
-                .unwrap_or_else(|payload| {
-                    crate::mask_events::CANDIDATE_PANIC
-                        .record_with(|| crate::mask_events::panic_payload(payload.as_ref()));
-                    maybe_dump_selected_llir("failed-filter-candidate", dyn_map, &llir_graph);
-                    if let Some(pre) = &pre_collapse {
-                        maybe_dump_selected_llir("failed-filter-precollapse", dyn_map, pre);
-                    }
-                    CandidateFilterResult::reject_with_display("candidate compile panicked")
-                });
-                if !filter_result.accepted {
-                    continue;
-                }
-                let filter_display = filter_result.display;
-
-                n_graphs += 1;
-                // Losers are the most expensive candidates to profile: a
-                // candidate whose running mean is already `factor ×` worse
-                // than the best can stop trialing — its partial mean is
-                // ranked normally and cannot win.
-                let early_stop = options
-                    .early_stop_factor
-                    .map(|factor| (best_metric.clone(), factor));
-                let profile_result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-                    let profile_start = std::time::Instant::now();
-                    let (rep_metric, rep_display) = runtime.profile_prepared_candidate(
-                        &llir_graph,
-                        &profile_dyn_map,
-                        options.trials,
-                        options.execution_timeout,
-                        early_stop,
-                        bucket_profile_context.as_ref().map(|bucket_context| {
-                            ProfileBucketContext {
-                                dim_buckets: &bucket_context.dim_buckets,
-                                bucket_indices: &bucket_context.bucket_indices,
-                                representative_dyn_map: &bucket_context.representative_dyn_map,
-                            }
-                        }),
-                    );
-                    let timed_out = candidate_timed_out(profile_start.elapsed());
-                    let invalid_profile = rep_display.starts_with("invalid ");
-                    (
-                        rep_metric,
-                        append_filter_display(rep_display, filter_display.as_deref()),
-                        timed_out,
-                        invalid_profile,
-                    )
-                }));
-
-                let (new_metric, display_metric) = match profile_result {
-                    Ok((metric, display, false, false)) => {
-                        generation_found_non_timeout = true;
-                        (R::aggregate_profile_metrics(&[metric]), display)
-                    }
-                    Ok((_, _, true, _)) | Err(_) => {
-                        // Timed out or panicked — redraw bars and skip.
-                        if search_log {
-                            for _ in 1..n_bar_lines {
-                                print!("\x1b[1A");
-                            }
-                            print!("\r\x1b[2K");
-                            render_bars(n_graphs, search_limit, bucket_progress);
-                            std::io::stdout().flush().unwrap();
-                        }
-                        continue;
-                    }
-                    Ok((_, _, false, true)) => {
-                        // Backend rejected this candidate during load/profile.
-                        if search_log {
-                            for _ in 1..n_bar_lines {
-                                print!("\x1b[1A");
-                            }
-                            print!("\r\x1b[2K");
-                            render_bars(n_graphs, search_limit, bucket_progress);
-                            std::io::stdout().flush().unwrap();
-                        }
-                        continue;
-                    }
-                };
-
-                let rank = ranked_candidates
-                    .iter()
-                    .position(|(metric, _)| {
-                        new_metric
-                            .partial_cmp(metric)
-                            .is_some_and(|ordering| ordering == std::cmp::Ordering::Less)
-                    })
-                    .unwrap_or(ranked_candidates.len());
-                ranked_candidates.insert(rank, (new_metric.clone(), genome.clone()));
-
-                // Update parents list (keep top-N for next generation)
-                let dominated_by_all = parents.len() >= options.keep_best
-                    && !parents.last().unwrap().0.gt(&new_metric);
-                if !dominated_by_all {
-                    let pos = parents
-                        .iter()
-                        .position(|(m, _)| {
-                            new_metric
-                                .partial_cmp(m)
-                                .is_some_and(|o| o == std::cmp::Ordering::Less)
-                        })
-                        .unwrap_or(parents.len());
-                    parents.insert(pos, (new_metric.clone(), genome.clone()));
-                    if parents.len() > options.keep_best {
-                        parents.truncate(options.keep_best);
-                    }
-                }
-
-                log_candidate_ops(&llir_graph, &format!("cand={n_graphs} {display_metric}"));
-                let new_best = best_metric.gt(&new_metric);
-                if new_best {
-                    generation_found_new_best = true;
-                    best_metric = new_metric;
-                    log_best_llir(
-                        &llir_graph,
-                        &format!("candidate={n_graphs} {display_metric}"),
-                    );
-                }
-
-                let msg = if new_best {
-                    slower_since_faster = 0;
-                    format!("   {:>6} {display_metric}", "Faster".green().bold())
-                } else {
-                    slower_since_faster += 1;
-                    format!("   {:>6} x{slower_since_faster}", "Slower".yellow().bold())
-                };
-                if search_log {
-                    // Move from the last progress bar to the first one. A
-                    // slower result has one additional transient line above
-                    // the bars; another slower result replaces it in place,
-                    // while a faster result is appended beneath it.
-                    for _ in 1..n_bar_lines {
-                        print!("\x1b[1A");
-                    }
-                    if slower_line_visible && !new_best {
-                        print!("\x1b[1A");
-                    }
-                    print!("\r\x1b[2K");
-                    println!("{msg}");
-                    slower_line_visible = !new_best;
-                }
-                if search_log {
-                    render_bars(n_graphs, search_limit, bucket_progress);
-                    std::io::stdout().flush().unwrap();
-                }
-            }
-
-            if generation_found_new_best {
-                stagnant_generations = 0;
-            } else {
-                stagnant_generations += 1;
-            }
-            // Every other stagnant generation past the threshold explores
-            // from fresh random genomes instead of the converged parents.
-            let stagnation_resample = options.restart_stagnation > 0
-                && stagnant_generations >= options.restart_stagnation
-                && stagnant_generations % 2 == 0;
-            resample_generation = !generation_found_non_timeout || stagnation_resample;
-        }
-
-        // Clear progress bars
-        if search_log {
-            for _ in 1..n_bar_lines {
-                print!("\x1b[1A");
-            }
-            print!("\r");
-            for _ in 0..n_bar_lines {
-                println!("\x1b[2K");
-            }
-            for _ in 0..n_bar_lines {
-                print!("\x1b[1A");
-            }
-            print!("\r");
-            std::io::stdout().flush().unwrap();
-        }
-
-        if search_log {
-            println!(
-                "   {:>6}  in {}",
-                "Searched".green().bold(),
-                pretty_duration::pretty_duration(&start.elapsed(), None)
-            );
-        }
-        ranked_candidates
-    }
-
-    /// Lazily materialize individually viable final LLIRs until `target` is
-    /// available. Ranked genomes remain compact e-graph choices; full graphs
-    /// are only retained when aggregate bucket backtracking actually reaches
-    /// them.
-    #[allow(clippy::too_many_arguments)]
-    fn ensure_finalist<R: Runtime + 'static>(
-        &self,
-        runtime: &mut R,
-        candidates: &mut LazyFinalists<R::ProfileMetric>,
-        target: usize,
-        options: &CompileOptions,
-        dyn_map: &DynMap,
-        bucket_profile_context: Option<&SearchProfileBucketContext>,
-        egraph_index: usize,
-        search_started_at: std::time::Instant,
-    ) -> bool {
-        let egraph = &self.egraphs[egraph_index];
-        let ops = self.ops.as_ref().unwrap();
-        let mut llir_extractor = LlirExtractor::new(egraph, ops);
-        let search_log = options.search_log_enabled();
-        let dump_pre_unroll = std::env::var_os("LLIR_DUMP_PRE_UNROLL").is_some();
-        let final_filter_dyn_map = if bucket_profile_context.is_some() {
-            dyn_map.clone()
-        } else {
-            let mut profile_dyn_map = dyn_map.clone();
-            for (&dim, &value) in &options.profile_dims {
-                profile_dyn_map.insert(dim, value);
-            }
-            profile_dyn_map
-        };
-
-        while candidates.finalists.len() <= target {
-            if candidates.next_ranked >= candidates.ranked.len() {
-                return false;
-            }
-            // Always give the fastest profiled genome one finalization attempt.
-            // Later fallbacks respect the overall search budget.
-            if candidates.next_ranked > 0
-                && search_started_at.elapsed() >= options.search_time_limit
-            {
-                candidates.stopped_reason = Some(format!(
-                    "search time limit expired before finalizing ranked candidate {}",
-                    candidates.next_ranked + 1
-                ));
-                return false;
-            }
-
-            let finalist_started_at = std::time::Instant::now();
-            let (metric, genome) = &candidates.ranked[candidates.next_ranked];
-            let metric = metric.clone();
-            let genome = genome.clone();
-            candidates.next_ranked += 1;
-
-            let final_graph_result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-                let packed = llir_extractor.extract_indexed_packed(&genome, &self.custom_ops, None);
-                let pre_unroll = dump_pre_unroll.then(|| packed.to_stable());
-                let stitched = unroll_packed_llir(packed);
-                (pre_unroll, stitched)
-            }));
-            if let Err(payload) = &final_graph_result {
-                crate::mask_events::CANDIDATE_PANIC
-                    .record_with(|| crate::mask_events::panic_payload(payload.as_ref()));
-            }
-            let Ok((pre_unroll, stitched)) = final_graph_result else {
-                candidates.rejections += 1;
-                candidates.last_rejection =
-                    Some("final extraction or loop unroll panicked".to_string());
-                if search_log {
-                    println!(
-                        "   {:>6}  finalist reject ranked #{}: final extraction or loop unroll panicked",
-                        "Search".yellow().bold(),
-                        candidates.next_ranked,
-                    );
-                }
-                continue;
-            };
-
-            runtime.clear_intermediate_buffers();
-            let filter_result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-                self.candidate_filter_result(
-                    runtime,
-                    &stitched,
-                    &final_filter_dyn_map,
-                    options,
-                    bucket_profile_context,
-                )
-            }))
-            .unwrap_or_else(|_| {
-                CandidateFilterResult::reject_with_display("final candidate filter panicked")
-            });
-
-            if options
-                .candidate_timeout
-                .is_some_and(|timeout| finalist_started_at.elapsed() >= timeout)
-            {
-                candidates.rejections += 1;
-                candidates.last_rejection = Some(format!(
-                    "candidate timeout expired while finalizing ranked candidate {}",
-                    candidates.next_ranked
-                ));
-                if search_log {
-                    println!(
-                        "   {:>6}  finalist reject ranked #{}: finalization timeout after {:?}",
-                        "Search".yellow().bold(),
-                        candidates.next_ranked,
-                        finalist_started_at.elapsed(),
-                    );
-                }
-                continue;
-            }
-            if !filter_result.accepted {
-                candidates.rejections += 1;
-                candidates.last_rejection = filter_result.display;
-                crate::mask_events::FINALIST_REJECT.record_with(|| {
-                    candidates
-                        .last_rejection
-                        .clone()
-                        .unwrap_or_else(|| "(no reason)".to_string())
-                });
-                if search_log {
-                    println!(
-                        "   {:>6}  finalist reject ranked #{}: {}",
-                        "Search".yellow().bold(),
-                        candidates.next_ranked,
-                        candidates
-                            .last_rejection
-                            .as_deref()
-                            .unwrap_or("(no reason)"),
-                    );
-                }
-                continue;
-            }
-
-            // Falling past rank #1 silently substitutes a slower-profiled
-            // graph; make the substitution visible.
-            if search_log && candidates.next_ranked > 1 {
-                println!(
-                    "   {:>6}  finalist fallback: loading ranked #{} after {} rejection(s)",
-                    "Search".yellow().bold(),
-                    candidates.next_ranked,
-                    candidates.rejections,
-                );
-            }
-            candidates.finalists.push(Finalist {
-                metric,
-                pre_unroll,
-                llir: stitched,
-            });
-        }
-        true
-    }
-
-    fn no_finalist_message<M>(candidates: &LazyFinalists<M>) -> String {
-        if let Some(stopped_reason) = &candidates.stopped_reason {
-            return format!(
-                "no viable final graph after {} hard-filter rejections: {stopped_reason}",
-                candidates.rejections
-            );
-        }
-        format!(
-            "no viable final graph after hard-filtering {} profiled candidates: {}",
-            candidates.rejections,
-            candidates
-                .last_rejection
-                .as_deref()
-                .unwrap_or("no rejection reason")
-        )
-    }
-
-    fn dump_selected_finalist<M>(
-        finalist: &Finalist<M>,
-        dyn_map: &DynMap,
-        bucket_progress: Option<(usize, usize)>,
-    ) {
-        if let Some(pre_unroll) = &finalist.pre_unroll {
-            let dump_label = bucket_progress
-                .map(|(bucket_idx, n_buckets)| {
-                    format!("pre-unroll-bucket-{:02}-of-{n_buckets:02}", bucket_idx + 1)
-                })
-                .unwrap_or_else(|| "pre-unroll-single".to_string());
-            maybe_dump_selected_llir(&dump_label, dyn_map, pre_unroll);
-        }
-        let dump_label = bucket_progress
-            .map(|(bucket_idx, n_buckets)| {
-                format!("bucket-{:02}-of-{n_buckets:02}", bucket_idx + 1)
-            })
-            .unwrap_or_else(|| "single".to_string());
-        maybe_dump_selected_llir(&dump_label, dyn_map, &finalist.llir);
-    }
-
-    fn candidate_filter_result<R: Runtime + 'static>(
-        &self,
-        runtime: &mut R,
-        llir_graph: &LLIRGraph,
-        dyn_map: &DynMap,
-        search_options: &CompileOptions,
-        bucket_profile_context: Option<&SearchProfileBucketContext>,
-    ) -> CandidateFilterResult {
-        runtime.filter_llir_candidate(
-            llir_graph,
-            CandidateFilterContext {
-                search_options,
-                dyn_map,
-                bucket_context: bucket_profile_context.map(|bucket_context| ProfileBucketContext {
-                    dim_buckets: &bucket_context.dim_buckets,
-                    bucket_indices: &bucket_context.bucket_indices,
-                    representative_dyn_map: &bucket_context.representative_dyn_map,
-                }),
-            },
-        )
-    }
-
-    fn prepare_profile_candidate<R: Runtime + 'static>(
-        &self,
-        runtime: &mut R,
-        llir_graph: &LLIRGraph,
-        dyn_map: &DynMap,
-        search_options: &CompileOptions,
-        bucket_profile_context: Option<&SearchProfileBucketContext>,
-    ) -> CandidateFilterResult {
-        runtime.prepare_profile_candidate(
-            llir_graph,
-            CandidateFilterContext {
-                search_options,
-                dyn_map,
-                bucket_context: bucket_profile_context.map(|bucket_context| ProfileBucketContext {
-                    dim_buckets: &bucket_context.dim_buckets,
-                    bucket_indices: &bucket_context.bucket_indices,
-                    representative_dyn_map: &bucket_context.representative_dyn_map,
-                }),
-            },
-        )
+        runtime.compile(space, &self.dyn_map, &options, rng);
+        runtime
     }
 }
-
 impl Deref for Graph {
     type Target = HLIRGraph;
     fn deref(&self) -> &Self::Target {
@@ -3156,14 +1819,6 @@ fn stable_toposort_by_node_index(graph: &HLIRGraph) -> Option<Vec<NodeIndex>> {
     }
 
     (ordered.len() == graph.node_count()).then_some(ordered)
-}
-
-fn append_filter_display(display: String, filter_display: Option<&str>) -> String {
-    if let Some(filter_display) = filter_display.filter(|s| !s.is_empty()) {
-        format!("{display} | {filter_display}")
-    } else {
-        display
-    }
 }
 
 struct RollingHash64 {
@@ -3609,1171 +2264,12 @@ fn grow_rolling_candidate(
     }
 }
 
-/// Append one line per profiled candidate (op-type histogram + metric) to
-/// the file named by `LUMINAL_CANDIDATE_OPS` — search-trajectory forensics
-/// for "was family X ever generated, and what did it measure".
-fn log_candidate_ops(llir: &LLIRGraph, tag: &str) {
-    static PATH: std::sync::OnceLock<Option<String>> = std::sync::OnceLock::new();
-    let Some(path) = PATH.get_or_init(|| std::env::var("LUMINAL_CANDIDATE_OPS").ok()) else {
-        return;
-    };
-    let mut counts: std::collections::BTreeMap<String, usize> = std::collections::BTreeMap::new();
-    for op in llir.node_weights() {
-        let debug = format!("{op:?}");
-        let name = debug
-            .split(['{', '(', ' ', ')'])
-            .find(|s| !s.is_empty() && *s != "LLIROp" && *s != "DialectOp")
-            .unwrap_or("?")
-            .to_string();
-        *counts.entry(name).or_default() += 1;
-    }
-    let line = format!(
-        "{tag} | {}\n",
-        counts
-            .iter()
-            .map(|(k, v)| format!("{k}:{v}"))
-            .collect::<Vec<_>>()
-            .join(",")
-    );
-    use std::io::Write;
-    if let Ok(mut f) = std::fs::OpenOptions::new()
-        .create(true)
-        .append(true)
-        .open(path)
-    {
-        let _ = f.write_all(line.as_bytes());
-    }
-}
-
-/// Expand all loop-region markers in an LLIR graph into fully unrolled bodies.
-///
-/// Reads `LoopStart` / `LoopEnd` / `LoopInput` / `LoopOutput` metadata placed
-/// by the auto-roll prepass, clones the loop body `iters-1` additional times,
-/// threads loop-carried state between clones, routes per-iteration inputs and
-/// per-iteration outputs, and removes the four marker op types.
-///
-/// Incoming-edge ORDER is preserved for every affected node — ops read their
-/// inputs by edge-id order, so edges are rebuilt in position.
-/// When `LUMINAL_LOG_LLIR=1`, print a canonical, diffable dump of a
-/// candidate LLIR each time the search finds a new fastest graph. Nodes are
-/// numbered canonically (Kahn topological order with a deterministic
-/// tie-break on op text and canonical input ids), so two runs of an
-/// identical graph produce byte-identical output regardless of NodeIndex
-/// assignment — best-so-far graphs from different runs can be compared with
-/// plain `diff`.
-pub fn log_best_llir(llir: &LLIRGraph, context: &str) {
-    static ENABLED: std::sync::OnceLock<bool> = std::sync::OnceLock::new();
-    if !*ENABLED.get_or_init(|| std::env::var_os("LUMINAL_LOG_LLIR").is_some()) {
-        return;
-    }
-    use petgraph::visit::EdgeRef;
-    use std::collections::BTreeMap;
-
-    let mut indegree: FxHashMap<NodeIndex, usize> = llir
-        .node_indices()
-        .map(|n| (n, llir.edges_directed(n, Direction::Incoming).count()))
-        .collect();
-    // Ready nodes keyed by (op text, canonical input ids) for deterministic
-    // pops; true duplicates tie and are interchangeable.
-    let mut ready: BTreeMap<(String, Vec<usize>), Vec<NodeIndex>> = BTreeMap::new();
-    let mut canonical: FxHashMap<NodeIndex, usize> = FxHashMap::default();
-    let inputs_of = |n: NodeIndex, canonical: &FxHashMap<NodeIndex, usize>| -> Vec<usize> {
-        llir.edges_directed(n, Direction::Incoming)
-            .sorted_by_key(|e| e.id())
-            .map(|e| canonical.get(&e.source()).copied().unwrap_or(usize::MAX))
-            .collect()
-    };
-    for (&n, &d) in &indegree {
-        if d == 0 {
-            ready
-                .entry((format!("{:?}", llir[n]), Vec::new()))
-                .or_default()
-                .push(n);
-        }
-    }
-    let mut lines: Vec<String> = Vec::with_capacity(indegree.len());
-    while let Some((key, nodes)) = ready.pop_first() {
-        for n in nodes {
-            let id = lines.len();
-            canonical.insert(n, id);
-            let inputs = key
-                .1
-                .iter()
-                .map(|i| format!("n{i}"))
-                .collect::<Vec<_>>()
-                .join(",");
-            lines.push(format!("n{id}: {} <- [{inputs}]", key.0));
-            for succ in llir
-                .neighbors_directed(n, Direction::Outgoing)
-                .collect::<Vec<_>>()
-            {
-                let d = indegree.get_mut(&succ).unwrap();
-                *d -= 1;
-                if *d == 0 {
-                    ready
-                        .entry((format!("{:?}", llir[succ]), inputs_of(succ, &canonical)))
-                        .or_default()
-                        .push(succ);
-                }
-            }
-        }
-    }
-    println!("LLIR_BEST {context} nodes={}", lines.len());
-    for line in &lines {
-        println!("{line}");
-    }
-    println!("LLIR_BEST_END");
-}
-
-/// Marker nodes and per-slot metadata of one rolled loop region, grouped
-/// from the LLIR graph by `loop_id`.
-#[derive(Default)]
-struct LoopRegion {
-    /// slot_idx → LoopStart.
-    starts: std::collections::BTreeMap<usize, NodeIndex>,
-    /// slot_idx → LoopEnd.
-    ends: std::collections::BTreeMap<usize, NodeIndex>,
-    /// stream_id → LoopInput.
-    inputs: std::collections::BTreeMap<usize, NodeIndex>,
-    /// stream_id → LoopOutput. Each stream has one LoopOutput.
-    outputs: std::collections::BTreeMap<usize, NodeIndex>,
-    /// LoopOutputSelect NodeIndex → (stream_id, iter).
-    output_selects: FxHashMap<NodeIndex, (usize, usize)>,
-    iters: usize,
-    /// Every marker node of this region.
-    markers: FxHashSet<NodeIndex>,
-}
-
-fn collect_loop_regions(llir: &LLIRGraph) -> std::collections::BTreeMap<usize, LoopRegion> {
-    use crate::hlir::{LoopEnd, LoopInput, LoopOutput, LoopOutputSelect, LoopStart};
-
-    let mut regions: std::collections::BTreeMap<usize, LoopRegion> =
-        std::collections::BTreeMap::new();
-    for n in llir.node_indices() {
-        let op = &llir[n];
-        let loop_id = if let Some(ls) = op.to_op::<LoopStart>() {
-            let region = regions.entry(ls.loop_id).or_default();
-            region.iters = region.iters.max(ls.iters.to_usize().unwrap_or(1));
-            region.starts.insert(ls.slot_idx, n);
-            ls.loop_id
-        } else if let Some(le) = op.to_op::<LoopEnd>() {
-            regions
-                .entry(le.loop_id)
-                .or_default()
-                .ends
-                .insert(le.slot_idx, n);
-            le.loop_id
-        } else if let Some(li) = op.to_op::<LoopInput>() {
-            regions
-                .entry(li.loop_id)
-                .or_default()
-                .inputs
-                .insert(li.stream_id, n);
-            li.loop_id
-        } else if let Some(los) = op.to_op::<LoopOutputSelect>() {
-            regions
-                .entry(los.loop_id)
-                .or_default()
-                .output_selects
-                .insert(n, (los.stream_id, los.iter));
-            los.loop_id
-        } else if let Some(lo) = op.to_op::<LoopOutput>() {
-            regions
-                .entry(lo.loop_id)
-                .or_default()
-                .outputs
-                .insert(lo.stream_id, n);
-            lo.loop_id
-        } else {
-            continue;
-        };
-        regions.entry(loop_id).or_default().markers.insert(n);
-    }
-    regions
-}
-
-/// Forward-reachable body of one region: successors of its entry markers,
-/// stopping at `Output` ops and at any loop marker of any region. Also
-/// reports whether a marker belonging to a *different* region was reached —
-/// i.e. that region is nested inside this one.
-fn loop_region_body(
-    llir: &LLIRGraph,
-    region: &LoopRegion,
-    marker_owner: &FxHashMap<NodeIndex, usize>,
-    self_id: usize,
-) -> (FxHashSet<NodeIndex>, std::collections::BTreeSet<usize>) {
-    use crate::hlir::Output;
-
-    let mut body_nodes: FxHashSet<NodeIndex> = FxHashSet::default();
-    let mut foreign: std::collections::BTreeSet<usize> = std::collections::BTreeSet::new();
-    let mut worklist: Vec<NodeIndex> = region
-        .starts
-        .values()
-        .chain(region.inputs.values())
-        .flat_map(|n| {
-            llir.neighbors_directed(*n, Direction::Outgoing)
-                .collect::<Vec<_>>()
-        })
-        .collect();
-    while let Some(n) = worklist.pop() {
-        if body_nodes.contains(&n) {
-            continue;
-        }
-        if let Some(&owner) = marker_owner.get(&n) {
-            if owner != self_id {
-                foreign.insert(owner);
-            }
-            continue;
-        }
-        if llir[n].to_op::<Output>().is_some() {
-            continue;
-        }
-        body_nodes.insert(n);
-        for succ in llir
-            .neighbors_directed(n, Direction::Outgoing)
-            .collect::<Vec<_>>()
-        {
-            worklist.push(succ);
-        }
-    }
-    (body_nodes, foreign)
-}
-
-/// Remove and return an innermost loop region — one whose body contains no
-/// other region's markers — together with that body. Marker ops are unique,
-/// so a repeated occurrence can never contain one and regions are always
-/// strictly nested or disjoint.
-fn take_innermost_region(llir: &LLIRGraph) -> Option<(LoopRegion, FxHashSet<NodeIndex>)> {
-    let mut regions = collect_loop_regions(llir);
-    if regions.is_empty() {
-        return None;
-    }
-    let marker_owner: FxHashMap<NodeIndex, usize> = regions
-        .iter()
-        .flat_map(|(&id, region)| region.markers.iter().map(move |&n| (n, id)))
-        .collect();
-    let (id, body_nodes) = regions
-        .iter()
-        .find_map(|(&id, region)| {
-            let (body_nodes, foreign) = loop_region_body(llir, region, &marker_owner, id);
-            foreign.is_empty().then_some((id, body_nodes))
-        })
-        .unwrap_or_else(|| {
-            let contains: Vec<String> = regions
-                .iter()
-                .map(|(&id, region)| {
-                    let (body_nodes, foreign) = loop_region_body(llir, region, &marker_owner, id);
-                    // Re-walk with parent tracking to show the exact bridge
-                    // edges into foreign markers.
-                    let mut bridges: Vec<String> = Vec::new();
-                    let mut seen: FxHashSet<NodeIndex> = FxHashSet::default();
-                    let mut worklist: Vec<(NodeIndex, Option<NodeIndex>)> = region
-                        .starts
-                        .values()
-                        .chain(region.inputs.values())
-                        .flat_map(|m| {
-                            llir.neighbors_directed(*m, Direction::Outgoing)
-                                .map(|s| (s, Some(*m)))
-                                .collect::<Vec<_>>()
-                        })
-                        .collect();
-                    while let Some((n, pred)) = worklist.pop() {
-                        if seen.contains(&n) {
-                            continue;
-                        }
-                        if let Some(&owner) = marker_owner.get(&n) {
-                            if owner != id && bridges.len() < 4 {
-                                let pred_desc = pred
-                                    .map(|p| format!("{:?}", llir[p]))
-                                    .unwrap_or_else(|| "<entry>".to_string());
-                                bridges.push(format!(
-                                    "      {} -> {:?}",
-                                    &pred_desc[..pred_desc.len().min(160)],
-                                    llir[n]
-                                ));
-                            }
-                            continue;
-                        }
-                        seen.insert(n);
-                        for succ in llir
-                            .neighbors_directed(n, Direction::Outgoing)
-                            .collect::<Vec<_>>()
-                        {
-                            worklist.push((succ, Some(n)));
-                        }
-                    }
-                    format!(
-                        "loop {id}: body={} starts={} inputs={} reaches markers of {foreign:?}\n{}",
-                        body_nodes.len(),
-                        region.starts.len(),
-                        region.inputs.len(),
-                        bridges.join("\n"),
-                    )
-                })
-                .collect();
-            panic!(
-                "loop regions must be strictly nested or disjoint; none is innermost:\n  {}",
-                contains.join("\n  ")
-            )
-        });
-    Some((regions.remove(&id).unwrap(), body_nodes))
-}
-
-/// Inline every iteration-invariant input marker: rewire its consumers to
-/// its single shared source and delete it. `LoopInputStatic` is invariant by
-/// definition (one source), and a `LoopInput` whose per-iteration sources are
-/// all the same node is invariant in fact. Egglog unions invariant markers
-/// with their source value, which lets extraction elect a marker node as the
-/// representative of a value class consumed far outside its region (an inner
-/// region's invariant input is often exactly an enclosing region's value).
-/// Region walks and rewiring must never see that cross-region aliasing, so
-/// both `unroll_loops_in_llir` and `collapse_loops_to_first_iter` inline them
-/// before touching any region.
-fn inline_static_loop_inputs(llir: &mut LLIRGraph) {
-    use crate::hlir::{LoopInput, LoopInputStatic};
-    use petgraph::visit::EdgeRef;
-
-    // One marker at a time with live edge reads: invariant markers chain
-    // (an inner region's invariant input is often an enclosing region's
-    // marker), so a source captured up front can be deleted before its
-    // dependent marker is processed.
-    while let Some((marker, source)) = llir.node_indices().find_map(|n| {
-        if llir[n].to_op::<LoopInputStatic>().is_none() && llir[n].to_op::<LoopInput>().is_none() {
-            return None;
-        }
-        let mut sources = llir.neighbors_directed(n, Direction::Incoming);
-        let first = sources.next()?;
-        sources.all(|s| s == first).then_some((n, first))
-    }) {
-        // Per-edge remove+add to keep each consumer's edge-id ordering via
-        // LIFO reuse — the runtime reads inputs sorted by edge id.
-        let consumers: Vec<(petgraph::graph::EdgeIndex, NodeIndex)> = llir
-            .edges_directed(marker, Direction::Outgoing)
-            .sorted_by_key(|e| e.id())
-            .map(|e| (e.id(), e.target()))
-            .collect();
-        for (eid, consumer) in consumers {
-            llir.remove_edge(eid);
-            llir.add_edge(source, consumer, ());
-        }
-        llir.remove_node(marker);
-        crate::mask_events::INVARIANT_MARKER_INLINED.record();
-    }
-}
-
-#[derive(Clone)]
-struct MaterializeLoopRegion {
-    iters: usize,
-}
-
-#[derive(Clone)]
-enum MaterializeLoopMarker {
-    Start {
-        region: usize,
-        initial: usize,
-        body_producer: usize,
-    },
-    End {
-        region: usize,
-        body_producer: usize,
-    },
-    Input {
-        region: usize,
-        sources: Vec<usize>,
-    },
-    StaticInput {
-        source: usize,
-    },
-    Output {
-        region: usize,
-        body_producer: usize,
-    },
-    OutputSelect {
-        region: usize,
-        iter: usize,
-        body_producer: usize,
-    },
-}
-
-struct MaterializeContextLayout {
-    membership: u128,
-    /// Region-local mixed-radix multipliers. Unlike the previous global
-    /// multiplier, this includes only loops that actually contain the node.
-    region_multipliers: Vec<(usize, usize)>,
-    count: usize,
-}
-
-impl MaterializeContextLayout {
-    fn new(mut membership: u128, regions: &[MaterializeLoopRegion]) -> Option<Self> {
-        let original_membership = membership;
-        let mut region_multipliers = Vec::with_capacity(membership.count_ones() as usize);
-        let mut count = 1usize;
-        while membership != 0 {
-            let region = membership.trailing_zeros() as usize;
-            region_multipliers.push((region, count));
-            count = count.checked_mul(regions[region].iters)?;
-            membership &= membership - 1;
-        }
-        Some(Self {
-            membership: original_membership,
-            region_multipliers,
-            count,
-        })
-    }
-
-    fn decode(&self, context: usize, regions: &[MaterializeLoopRegion], iterations: &mut [usize]) {
-        debug_assert!(context < self.count);
-        for &(region, multiplier) in &self.region_multipliers {
-            iterations[region] = (context / multiplier) % regions[region].iters;
-        }
-    }
-
-    fn encode(
-        &self,
-        assigned_membership: u128,
-        regions: &[MaterializeLoopRegion],
-        iterations: &[usize],
-    ) -> Option<usize> {
-        if self.membership & !assigned_membership != 0 {
-            return None;
-        }
-        let mut context = 0usize;
-        for &(region, multiplier) in &self.region_multipliers {
-            let iter = iterations[region];
-            if iter >= regions[region].iters {
-                return None;
-            }
-            context = context.checked_add(iter.checked_mul(multiplier)?)?;
-        }
-        Some(context)
-    }
-}
-
-struct MaterializeAdjacency {
-    offsets: Vec<usize>,
-    neighbors: Vec<usize>,
-}
-
-impl MaterializeAdjacency {
-    fn incoming(llir: &LLIRGraph, node_bound: usize) -> Self {
-        let mut offsets = vec![0usize; node_bound + 1];
-        for edge in llir.edge_indices() {
-            let (_, destination) = llir.edge_endpoints(edge).unwrap();
-            offsets[destination.index() + 1] += 1;
-        }
-        for index in 1..offsets.len() {
-            offsets[index] += offsets[index - 1];
-        }
-        let mut cursor = offsets[..node_bound].to_vec();
-        let mut neighbors = vec![usize::MAX; llir.edge_count()];
-        // StableGraph edge indices iterate in ascending slot order. Filling
-        // each destination's range in that order preserves positional inputs
-        // without sorting or allocating one Vec per node.
-        for edge in llir.edge_indices() {
-            let (source, destination) = llir.edge_endpoints(edge).unwrap();
-            neighbors[cursor[destination.index()]] = source.index();
-            cursor[destination.index()] += 1;
-        }
-        Self { offsets, neighbors }
-    }
-
-    fn outgoing(llir: &LLIRGraph, node_bound: usize) -> Self {
-        let mut offsets = vec![0usize; node_bound + 1];
-        for edge in llir.edge_indices() {
-            let (source, _) = llir.edge_endpoints(edge).unwrap();
-            offsets[source.index() + 1] += 1;
-        }
-        for index in 1..offsets.len() {
-            offsets[index] += offsets[index - 1];
-        }
-        let mut cursor = offsets[..node_bound].to_vec();
-        let mut neighbors = vec![usize::MAX; llir.edge_count()];
-        for edge in llir.edge_indices() {
-            let (source, destination) = llir.edge_endpoints(edge).unwrap();
-            neighbors[cursor[source.index()]] = destination.index();
-            cursor[source.index()] += 1;
-        }
-        Self { offsets, neighbors }
-    }
-
-    fn get(&self, node: usize) -> &[usize] {
-        &self.neighbors[self.offsets[node]..self.offsets[node + 1]]
-    }
-}
-
-trait MaterializeLLIRView {
-    fn node_bound(&self) -> usize;
-    fn op(&self, node: usize) -> Option<&LLIROp>;
-    fn incoming(&self, node: usize) -> &[usize];
-    fn outgoing(&self, node: usize) -> &[usize];
-}
-
-struct StableMaterializeView<'a> {
-    graph: &'a LLIRGraph,
-    incoming: MaterializeAdjacency,
-    outgoing: MaterializeAdjacency,
-    node_bound: usize,
-}
-
-impl<'a> StableMaterializeView<'a> {
-    fn new(graph: &'a LLIRGraph) -> Self {
-        let node_bound = petgraph::visit::NodeIndexable::node_bound(graph);
-        Self {
-            graph,
-            incoming: MaterializeAdjacency::incoming(graph, node_bound),
-            outgoing: MaterializeAdjacency::outgoing(graph, node_bound),
-            node_bound,
-        }
-    }
-}
-
-impl MaterializeLLIRView for StableMaterializeView<'_> {
-    fn node_bound(&self) -> usize {
-        self.node_bound
-    }
-
-    fn op(&self, node: usize) -> Option<&LLIROp> {
-        self.graph.node_weight(NodeIndex::new(node))
-    }
-
-    fn incoming(&self, node: usize) -> &[usize] {
-        self.incoming.get(node)
-    }
-
-    fn outgoing(&self, node: usize) -> &[usize] {
-        self.outgoing.get(node)
-    }
-}
-
-impl MaterializeLLIRView for PackedLLIRGraph {
-    fn node_bound(&self) -> usize {
-        self.ops.len()
-    }
-
-    fn op(&self, node: usize) -> Option<&LLIROp> {
-        self.ops.get(node)
-    }
-
-    fn incoming(&self, node: usize) -> &[usize] {
-        &self.incoming_sources[self.incoming_offsets[node]..self.incoming_offsets[node + 1]]
-    }
-
-    fn outgoing(&self, node: usize) -> &[usize] {
-        &self.outgoing_targets[self.outgoing_offsets[node]..self.outgoing_offsets[node + 1]]
-    }
-}
-
-#[derive(Default)]
-struct MaterializeCollectedRegion {
-    starts: std::collections::BTreeMap<usize, usize>,
-    ends: std::collections::BTreeMap<usize, usize>,
-    inputs: std::collections::BTreeMap<usize, usize>,
-    outputs: std::collections::BTreeMap<usize, usize>,
-    output_selects: FxHashMap<usize, (usize, usize)>,
-    iters: usize,
-    markers: FxHashSet<usize>,
-}
-
-fn collect_materialize_loop_regions(
-    llir: &impl MaterializeLLIRView,
-) -> std::collections::BTreeMap<usize, MaterializeCollectedRegion> {
-    use crate::hlir::{LoopEnd, LoopInput, LoopOutput, LoopOutputSelect, LoopStart};
-
-    let mut regions: std::collections::BTreeMap<usize, MaterializeCollectedRegion> =
-        std::collections::BTreeMap::new();
-    for node in 0..llir.node_bound() {
-        let Some(op) = llir.op(node) else {
-            continue;
-        };
-        let loop_id = if let Some(marker) = op.to_op::<LoopStart>() {
-            let region = regions.entry(marker.loop_id).or_default();
-            region.iters = region.iters.max(marker.iters.to_usize().unwrap_or(1));
-            region.starts.insert(marker.slot_idx, node);
-            marker.loop_id
-        } else if let Some(marker) = op.to_op::<LoopEnd>() {
-            regions
-                .entry(marker.loop_id)
-                .or_default()
-                .ends
-                .insert(marker.slot_idx, node);
-            marker.loop_id
-        } else if let Some(marker) = op.to_op::<LoopInput>() {
-            regions
-                .entry(marker.loop_id)
-                .or_default()
-                .inputs
-                .insert(marker.stream_id, node);
-            marker.loop_id
-        } else if let Some(marker) = op.to_op::<LoopOutputSelect>() {
-            regions
-                .entry(marker.loop_id)
-                .or_default()
-                .output_selects
-                .insert(node, (marker.stream_id, marker.iter));
-            marker.loop_id
-        } else if let Some(marker) = op.to_op::<LoopOutput>() {
-            regions
-                .entry(marker.loop_id)
-                .or_default()
-                .outputs
-                .insert(marker.stream_id, node);
-            marker.loop_id
-        } else {
-            continue;
-        };
-        regions.entry(loop_id).or_default().markers.insert(node);
-    }
-    regions
-}
-
-/// Materialize a rolled LLIR into one fully-unrolled StableGraph without ever
-/// mutating a graph containing loop markers.  Each surviving `(node, loop
-/// context)` pair is allocated exactly once and its ordered incoming edges are
-/// emitted directly into the final graph.  This avoids StableGraph free-list
-/// edge reuse and therefore needs no repair/compaction pass.
-fn materialize_unrolled_view(llir: &impl MaterializeLLIRView) -> Option<LLIRGraph> {
-    use crate::hlir::{
-        LoopEnd, LoopInput, LoopInputStatic, LoopOutput, LoopOutputSelect, LoopStart, Output,
-    };
-    let node_bound = llir.node_bound();
-
-    let mut loop_regions = collect_materialize_loop_regions(llir);
-    let mut markers: Vec<Option<MaterializeLoopMarker>> = vec![None; node_bound];
-
-    // Invariant LoopInput variants are transparent and must not seed a loop
-    // body. This mirrors `inline_static_loop_inputs` before region discovery.
-    for (node, marker_slot) in markers.iter_mut().enumerate() {
-        let Some(op) = llir.op(node) else {
-            continue;
-        };
-        if op.to_op::<LoopInputStatic>().is_some() {
-            let &source = llir.incoming(node).first()?;
-            *marker_slot = Some(MaterializeLoopMarker::StaticInput { source });
-        } else if let Some(input) = op.to_op::<LoopInput>() {
-            let sources = llir.incoming(node);
-            if let Some(&first) = sources.first()
-                && sources.iter().all(|source| *source == first)
-            {
-                *marker_slot = Some(MaterializeLoopMarker::StaticInput { source: first });
-                if let Some(region) = loop_regions.get_mut(&input.loop_id) {
-                    region.inputs.remove(&input.stream_id);
-                    region.markers.remove(&node);
-                }
-            }
-        }
-    }
-
-    if loop_regions.len() > 128 {
-        return None;
-    }
-
-    let mut region_by_id: FxHashMap<usize, usize> = FxHashMap::default();
-    let mut regions = Vec::with_capacity(loop_regions.len());
-    for (&loop_id, region) in &loop_regions {
-        if region.iters <= 1 || region.starts.is_empty() {
-            return None;
-        }
-        region_by_id.insert(loop_id, regions.len());
-        regions.push(MaterializeLoopRegion {
-            iters: region.iters,
-        });
-    }
-
-    let mut marker_owner = vec![None; node_bound];
-    for (&loop_id, region) in &loop_regions {
-        let region_index = region_by_id[&loop_id];
-        for &marker in &region.markers {
-            marker_owner[marker] = Some(region_index);
-        }
-
-        for (&slot, &start) in &region.starts {
-            let &end = region.ends.get(&slot)?;
-            let &initial = llir.incoming(start).first()?;
-            let &body_producer = llir.incoming(end).first()?;
-            markers[start] = Some(MaterializeLoopMarker::Start {
-                region: region_index,
-                initial,
-                body_producer,
-            });
-            markers[end] = Some(MaterializeLoopMarker::End {
-                region: region_index,
-                body_producer,
-            });
-        }
-
-        for &input in region.inputs.values() {
-            let sources = llir.incoming(input).to_vec();
-            if sources.len() != region.iters {
-                return None;
-            }
-            markers[input] = Some(MaterializeLoopMarker::Input {
-                region: region_index,
-                sources,
-            });
-        }
-
-        let mut output_producers = FxHashMap::default();
-        for (&stream, &output) in &region.outputs {
-            let &body_producer = llir.incoming(output).first()?;
-            output_producers.insert(stream, body_producer);
-            markers[output] = Some(MaterializeLoopMarker::Output {
-                region: region_index,
-                body_producer,
-            });
-        }
-        for (&select, &(stream, iter)) in &region.output_selects {
-            if iter >= region.iters {
-                return None;
-            }
-            markers[select] = Some(MaterializeLoopMarker::OutputSelect {
-                region: region_index,
-                iter,
-                body_producer: *output_producers.get(&stream)?,
-            });
-        }
-    }
-
-    // Reject malformed marker graphs rather than silently materializing a
-    // marker as an executable op. Callers treat this as an invariant failure.
-    for (node, marker) in markers.iter().enumerate() {
-        let Some(op) = llir.op(node) else {
-            continue;
-        };
-        let is_marker = op.to_op::<LoopStart>().is_some()
-            || op.to_op::<LoopEnd>().is_some()
-            || op.to_op::<LoopInput>().is_some()
-            || op.to_op::<LoopInputStatic>().is_some()
-            || op.to_op::<LoopOutput>().is_some()
-            || op.to_op::<LoopOutputSelect>().is_some();
-        if is_marker && marker.is_none() {
-            return None;
-        }
-    }
-
-    // Region membership is a transitive scope: foreign (nested) markers are
-    // transparent, while this region's own exit markers are boundaries. This
-    // directly describes the nodes that must be cloned for each enclosing
-    // iteration.
-    let mut membership = vec![0u128; node_bound];
-    for (&loop_id, region) in &loop_regions {
-        let region_index = region_by_id[&loop_id];
-        let mut seen = vec![false; node_bound];
-        let mut worklist: Vec<usize> = region
-            .starts
-            .values()
-            .chain(region.inputs.values())
-            .flat_map(|marker| llir.outgoing(*marker).iter().copied())
-            .collect();
-        while let Some(node) = worklist.pop() {
-            if seen[node] {
-                continue;
-            }
-            seen[node] = true;
-            if markers[node].is_some() {
-                if marker_owner[node] != Some(region_index) {
-                    worklist.extend(llir.outgoing(node));
-                }
-                continue;
-            }
-            if llir.op(node)?.to_op::<Output>().is_some() {
-                continue;
-            }
-            membership[node] |= 1u128 << region_index;
-            worklist.extend(llir.outgoing(node));
-        }
-    }
-
-    // Nodes only need instances for the loops that contain them. Independent
-    // regions therefore remain independent instead of contributing to a
-    // graph-wide Cartesian product. A membership containing multiple regions
-    // represents genuine nesting, where the product is semantically required.
-    let mut contexts_by_membership: FxHashMap<u128, MaterializeContextLayout> =
-        FxHashMap::default();
-    contexts_by_membership.insert(0, MaterializeContextLayout::new(0, &regions)?);
-    let mut node_instance_offsets = vec![usize::MAX; node_bound];
-    let mut final_node_count = 0usize;
-    let mut final_edge_count = 0usize;
-    for node in 0..node_bound {
-        if llir.op(node).is_none() || markers[node].is_some() {
-            continue;
-        }
-        let context_count = match contexts_by_membership.entry(membership[node]) {
-            Entry::Occupied(entry) => entry.get().count,
-            Entry::Vacant(entry) => {
-                entry
-                    .insert(MaterializeContextLayout::new(membership[node], &regions)?)
-                    .count
-            }
-        };
-        node_instance_offsets[node] = final_node_count;
-        final_node_count = final_node_count.checked_add(context_count)?;
-        final_edge_count =
-            final_edge_count.checked_add(context_count.checked_mul(llir.incoming(node).len())?)?;
-    }
-
-    let mut materialized = LLIRGraph::with_capacity(final_node_count, final_edge_count);
-
-    for node in 0..node_bound {
-        let Some(op) = llir.op(node) else {
-            continue;
-        };
-        if markers[node].is_some() {
-            continue;
-        }
-        let layout = &contexts_by_membership[&membership[node]];
-        for context in 0..layout.count {
-            let materialized_node = materialized.add_node(op.clone());
-            debug_assert_eq!(
-                materialized_node.index(),
-                node_instance_offsets[node] + context
-            );
-        }
-    }
-
-    let mut context_iterations = vec![0usize; regions.len()];
-    for node in 0..node_bound {
-        if llir.op(node).is_none() || markers[node].is_some() {
-            continue;
-        }
-        let node_membership = membership[node];
-        let node_layout = &contexts_by_membership[&node_membership];
-        for context in 0..node_layout.count {
-            let target = NodeIndex::new(node_instance_offsets[node] + context);
-            for &original_source in llir.incoming(node) {
-                node_layout.decode(context, &regions, &mut context_iterations);
-                let mut assigned_membership = node_membership;
-                let mut source = original_source;
-                let mut marker_hops = 0usize;
-                while let Some(marker) = markers[source].as_ref() {
-                    marker_hops += 1;
-                    if marker_hops > node_bound {
-                        return None;
-                    }
-                    match marker {
-                        MaterializeLoopMarker::Start {
-                            region,
-                            initial,
-                            body_producer,
-                        } => {
-                            let region_bit = 1u128 << *region;
-                            if assigned_membership & region_bit == 0 {
-                                return None;
-                            }
-                            let iter = context_iterations[*region];
-                            if iter == 0 {
-                                source = *initial;
-                            } else {
-                                source = *body_producer;
-                                context_iterations[*region] = iter - 1;
-                            }
-                        }
-                        MaterializeLoopMarker::End {
-                            region,
-                            body_producer,
-                        } => {
-                            source = *body_producer;
-                            assigned_membership |= 1u128 << *region;
-                            context_iterations[*region] = regions[*region].iters - 1;
-                        }
-                        MaterializeLoopMarker::Input { region, sources } => {
-                            let region_bit = 1u128 << *region;
-                            if assigned_membership & region_bit == 0 {
-                                return None;
-                            }
-                            source = sources[context_iterations[*region]];
-                        }
-                        MaterializeLoopMarker::StaticInput {
-                            source: static_source,
-                        } => {
-                            source = *static_source;
-                        }
-                        MaterializeLoopMarker::Output {
-                            region,
-                            body_producer,
-                        } => {
-                            source = *body_producer;
-                            let region_bit = 1u128 << *region;
-                            if assigned_membership & region_bit == 0 {
-                                assigned_membership |= region_bit;
-                                context_iterations[*region] = 0;
-                            }
-                        }
-                        MaterializeLoopMarker::OutputSelect {
-                            region,
-                            iter,
-                            body_producer,
-                        } => {
-                            source = *body_producer;
-                            assigned_membership |= 1u128 << *region;
-                            context_iterations[*region] = *iter;
-                        }
-                    }
-                }
-                if llir.op(source).is_none() || node_instance_offsets[source] == usize::MAX {
-                    return None;
-                }
-                let source_layout = contexts_by_membership.get(&membership[source])?;
-                let source_context =
-                    source_layout.encode(assigned_membership, &regions, &context_iterations)?;
-                let materialized_source =
-                    NodeIndex::new(node_instance_offsets[source] + source_context);
-                materialized.add_edge(materialized_source, target, ());
-            }
-        }
-    }
-
-    Some(materialized)
-}
-
-fn materialize_unrolled_llir(llir: &LLIRGraph) -> Option<LLIRGraph> {
-    materialize_unrolled_view(&StableMaterializeView::new(llir))
-}
-
-/// Complete the indexed hot path by materializing its dense rolled graph
-/// directly into the public StableGraph representation expected by runtimes.
-pub(crate) fn unroll_packed_llir(llir: PackedLLIRGraph) -> LLIRGraph {
-    let started_at = std::time::Instant::now();
-    let rolled_nodes = llir.ops.len();
-    let graph = materialize_unrolled_view(&llir)
-        .expect("rolled LLIR must satisfy the loop materialization invariants");
-    if crate::egglog_utils::llir_profile_enabled() {
-        eprintln!(
-            "LLIR_UNROLL_PROFILE total_ms={:.3} rolled_nodes={} unrolled_nodes={} unrolled_edges={}",
-            started_at.elapsed().as_secs_f64() * 1e3,
-            rolled_nodes,
-            graph.node_count(),
-            graph.edge_count(),
-        );
-    }
-    graph
-}
-
-pub fn unroll_loops_in_llir(llir: &mut LLIRGraph) {
-    let started_at = std::time::Instant::now();
-    let rolled_nodes = llir.node_count();
-    *llir = materialize_unrolled_llir(llir)
-        .expect("rolled LLIR must satisfy the loop materialization invariants");
-    if crate::egglog_utils::llir_profile_enabled() {
-        eprintln!(
-            "LLIR_UNROLL_PROFILE total_ms={:.3} rolled_nodes={} unrolled_nodes={} unrolled_edges={}",
-            started_at.elapsed().as_secs_f64() * 1e3,
-            rolled_nodes,
-            llir.node_count(),
-            llir.edge_count(),
-        );
-    }
-}
-
-/// Collapse all loop markers in an LLIR graph down to a SINGLE iteration's
-/// body, with first-iteration inputs and outputs only. This is the cheap
-/// per-candidate form used by the genetic search — profiling one transformer
-/// block instead of N×block makes the search ~N× faster, and the relative
-/// cost of any extraction choice is preserved on the body shape.
-///
-/// LoopStart consumers re-route to the initial value, LoopInput consumers
-/// re-route to `sources[0]`, LoopEnd's post-loop consumers re-route to the body producer
-/// directly, and each `LoopOutput` is replaced with a single `Output { node:
-/// targets[0] }`. After collapse the LLIR has no marker ops left and contains
-/// exactly the iter-0 body plus the surrounding non-loop graph.
-pub fn collapse_loops_to_first_iter(llir: &mut LLIRGraph) {
-    inline_static_loop_inputs(llir);
-    // Innermost first: collapsing an inner region leaves its iter-0 body as
-    // plain nodes inside the enclosing region, which then collapses over them
-    // in turn.
-    while let Some((region, body)) = take_innermost_region(llir) {
-        if region.starts.is_empty() {
-            eprintln!(
-                "[loop-debug] collapse abandoned on degenerate region: ends={} inputs={} outputs={} selects={}",
-                region.ends.len(),
-                region.inputs.len(),
-                region.outputs.len(),
-                region.output_selects.len(),
-            );
-            return;
-        }
-        collapse_loop_region(llir, &region, &body);
-        let compacted = compact_llir_preserving_input_order(llir);
-        *llir = compacted;
-    }
-}
-
-fn collapse_loop_region(
-    llir: &mut LLIRGraph,
-    region: &LoopRegion,
-    body_nodes: &FxHashSet<NodeIndex>,
-) {
-    use petgraph::visit::EdgeRef;
-
-    let LoopRegion {
-        starts,
-        ends,
-        inputs,
-        outputs,
-        output_selects,
-        markers: loop_markers,
-        ..
-    } = region;
-
-    // Initial value per LoopStart, body producer per LoopEnd / LoopOutput.
-    let mut start_initial: FxHashMap<NodeIndex, NodeIndex> = FxHashMap::default();
-    for &start_node in starts.values() {
-        let initial = llir
-            .neighbors_directed(start_node, Direction::Incoming)
-            .next()
-            .expect("LoopStart must have an initial-value producer");
-        start_initial.insert(start_node, initial);
-    }
-    let mut input_first_source: FxHashMap<NodeIndex, NodeIndex> = FxHashMap::default();
-    for input_node in inputs.values() {
-        let first = llir
-            .edges_directed(*input_node, Direction::Incoming)
-            .sorted_by_key(|e| e.id())
-            .map(|e| e.source())
-            .next()
-            .expect("LoopInput must have at least one source");
-        input_first_source.insert(*input_node, first);
-    }
-
-    // Resolve a source reference to its iter-0 equivalent.
-    let resolve_src = |src: NodeIndex| -> NodeIndex {
-        if let Some(&initial) = start_initial.get(&src) {
-            initial
-        } else if let Some(&first) = input_first_source.get(&src) {
-            first
-        } else {
-            src
-        }
-    };
-
-    // Rewrite every body node's incoming edges. Per-edge remove+add to keep
-    // edge-id ordering via LIFO reuse — runtime reads inputs sorted by edge
-    // id so position must be preserved.
-    for &b in body_nodes {
-        let pairs: Vec<(NodeIndex, petgraph::graph::EdgeIndex)> = llir
-            .edges_directed(b, Direction::Incoming)
-            .sorted_by_key(|e| e.id())
-            .map(|e| (e.source(), e.id()))
-            .collect();
-        for (src, eid) in pairs {
-            let new_src = resolve_src(src);
-            llir.remove_edge(eid);
-            llir.add_edge(new_src, b, ());
-        }
-    }
-
-    // Per LoopOutput stream, find the body producer (its single incoming edge).
-    let mut output_body_producer: FxHashMap<usize, NodeIndex> = FxHashMap::default();
-    for (&stream_id, &output_node) in outputs {
-        let body_producer = llir
-            .neighbors_directed(output_node, Direction::Incoming)
-            .next()
-            .expect("LoopOutput missing body producer during rewire");
-        output_body_producer.insert(stream_id, body_producer);
-    }
-
-    // Post-loop consumers reading from LoopEnd / LoopOutputSelect must
-    // instead read from the body producer (iter-0's value) directly. In the
-    // collapsed form every Select(i) — regardless of i — re-routes to iter-0's
-    // body producer; iter > 0 Selects don't have a real value to forward, so
-    // they alias iter 0's. This keeps post-loop graph topology unchanged.
-    let mut marker_post_sub: FxHashMap<NodeIndex, NodeIndex> = FxHashMap::default();
-    for &end_node in ends.values() {
-        let body_producer = llir
-            .neighbors_directed(end_node, Direction::Incoming)
-            .next()
-            .expect("LoopEnd missing body producer during rewire");
-        marker_post_sub.insert(end_node, body_producer);
-    }
-    for (&select_node, &(stream_id, _)) in output_selects {
-        if let Some(&body_producer) = output_body_producer.get(&stream_id) {
-            marker_post_sub.insert(select_node, body_producer);
-        }
-    }
-    // Entry markers can also have consumers outside the walked body:
-    // extraction may elect a marker as the representative of a value class
-    // read anywhere in the graph. Resolve them to their iter-0 values —
-    // never leave a consumer pointing at a marker about to be removed.
-    for (&input_node, &first) in &input_first_source {
-        marker_post_sub.insert(input_node, first);
-    }
-    for (&start_node, &initial) in &start_initial {
-        marker_post_sub.insert(start_node, initial);
-    }
-    let post_loop_consumers: FxHashSet<NodeIndex> = loop_markers
-        .iter()
-        .flat_map(|n| {
-            llir.neighbors_directed(*n, Direction::Outgoing)
-                .collect::<Vec<_>>()
-        })
-        .filter(|n| !loop_markers.contains(n) && !body_nodes.contains(n))
-        .collect();
-    for &consumer in &post_loop_consumers {
-        let pairs: Vec<(NodeIndex, petgraph::graph::EdgeIndex)> = llir
-            .edges_directed(consumer, Direction::Incoming)
-            .sorted_by_key(|e| e.id())
-            .map(|e| (e.source(), e.id()))
-            .collect();
-        for (src, eid) in pairs {
-            let new_src = marker_post_sub.get(&src).copied().unwrap_or(src);
-            llir.remove_edge(eid);
-            llir.add_edge(new_src, consumer, ());
-        }
-    }
-
-    for &n in loop_markers {
-        llir.remove_node(n);
-    }
-}
-
-/// Rebuild an LLIR graph into a fresh StableGraph, copying nodes and edges
-/// such that edge IDs are sequential in the insertion order we choose
-/// (per-node incoming edges in their original edge-id order). This erases
-/// any free-list reuse artifacts from prior `remove_edge` / `remove_node`
-/// calls.
-fn compact_llir_preserving_input_order(old: &LLIRGraph) -> LLIRGraph {
-    use petgraph::visit::EdgeRef;
-    let mut new_graph = LLIRGraph::default();
-    let mut old_to_new: FxHashMap<NodeIndex, NodeIndex> = FxHashMap::default();
-    // Topo sort to add nodes in a deterministic order. If the graph has
-    // cycles (shouldn't for LLIR), fall back to node_indices order.
-    let topo = match petgraph::algo::toposort(old, None) {
-        Ok(v) => v,
-        Err(_) => old.node_indices().collect(),
-    };
-    for n in &topo {
-        let new_n = new_graph.add_node(old[*n].clone());
-        old_to_new.insert(*n, new_n);
-    }
-    // Add edges in topo order, per-node incoming sorted by old edge id.
-    // This reassigns new edge indices sequentially so sort-by-id matches
-    // the intended input position.
-    for n in &topo {
-        let incoming: Vec<NodeIndex> = old
-            .edges_directed(*n, Direction::Incoming)
-            .sorted_by_key(|e| e.id())
-            .map(|e| e.source())
-            .collect();
-        for src in incoming {
-            if let (Some(&new_src), Some(&new_dst)) = (old_to_new.get(&src), old_to_new.get(n)) {
-                new_graph.add_edge(new_src, new_dst, ());
-            }
-        }
-    }
-    new_graph
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::egglog_utils::hash_egglog_normalized;
-    use crate::egglog_utils::{
-        api::{Rule, SortDef, sort},
-        base::OP_KIND,
-    };
-    use crate::hlir::{
-        Input, LoopEnd, LoopInput, LoopStart, Output, ReferenceData, ReferenceOp, Sin,
-    };
-    use rand::SeedableRng;
+    use crate::hlir::{Input, LoopEnd, LoopInput, LoopStart, Output, ReferenceOp, Sin};
+    use crate::search::unroll::materialize_unrolled_llir;
 
     // A rolling candidate is only collapsible if every non-state boundary input
     // is fed from OUTSIDE the candidate's occurrences. A non-state input produced
@@ -4808,7 +2304,6 @@ mod tests {
         assert!(!candidate_is_rollable(&[occ(0, 100)], &[]));
     }
     use crate::tests::{assert_close, random_vec};
-    use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
 
     #[test]
     fn materialize_many_disjoint_loops_without_a_global_cartesian_product() {
@@ -4855,1181 +2350,6 @@ mod tests {
             materialized
                 .node_weights()
                 .all(|op| { op.to_op::<LoopStart>().is_none() && op.to_op::<LoopEnd>().is_none() })
-        );
-    }
-
-    static SEARCH_DIM_LATE_PASS_CALLED: AtomicBool = AtomicBool::new(false);
-    static SEARCH_DIM_LATE_PASS_SAW_C: AtomicBool = AtomicBool::new(false);
-
-    #[derive(Default)]
-    struct SearchDimRecordingRuntime {
-        profile_dyn_maps: Vec<DynMap>,
-        bucket_representative_dyn_maps: Vec<DynMap>,
-    }
-
-    impl Runtime for SearchDimRecordingRuntime {
-        type Ops = ();
-        type CompileArg = ();
-        type ExecReturn = ();
-        type ProfileMetric = usize;
-
-        fn late_egglog_passes(
-            _: &[Arc<Box<dyn EgglogOp>>],
-            _: &CompileOptions,
-            dyn_map: &DynMap,
-        ) -> Vec<crate::egglog_utils::LateEgglogPass> {
-            SEARCH_DIM_LATE_PASS_CALLED.store(true, Ordering::SeqCst);
-            SEARCH_DIM_LATE_PASS_SAW_C.store(dyn_map.contains_key(&sym("c")), Ordering::SeqCst);
-            vec![]
-        }
-
-        fn initialize(_: Self::CompileArg) -> Self {
-            Self::default()
-        }
-
-        fn load_llir(&mut self, _: &LLIRGraph) {}
-
-        fn load_llir_buckets(
-            &mut self,
-            _: &FxHashMap<Symbol, Vec<DimBucket>>,
-            bucket_llirs: &[BucketLLIR],
-        ) {
-            self.bucket_representative_dyn_maps = bucket_llirs
-                .iter()
-                .map(|(_, representative_dyn_map, _)| representative_dyn_map.clone())
-                .collect();
-        }
-
-        fn execute(&mut self, _: &DynMap) -> Self::ExecReturn {}
-
-        fn profile(
-            &mut self,
-            _: &LLIRGraph,
-            dyn_map: &DynMap,
-            _: usize,
-            _: Option<std::time::Duration>,
-            _: Option<(Self::ProfileMetric, f64)>,
-        ) -> (Self::ProfileMetric, String) {
-            self.profile_dyn_maps.push(dyn_map.clone());
-            (0, "0 ms".to_string())
-        }
-    }
-
-    #[derive(Default)]
-    struct TestFilterRuntime {
-        reject_candidates: bool,
-    }
-
-    impl Runtime for TestFilterRuntime {
-        type Ops = ();
-        type CompileArg = ();
-        type ExecReturn = ();
-        type ProfileMetric = usize;
-
-        fn initialize(_: Self::CompileArg) -> Self {
-            Self::default()
-        }
-
-        fn load_llir(&mut self, _: &LLIRGraph) {}
-
-        fn execute(&mut self, _: &DynMap) -> Self::ExecReturn {}
-
-        fn profile(
-            &mut self,
-            _: &LLIRGraph,
-            _: &DynMap,
-            _: usize,
-            _: Option<std::time::Duration>,
-            _: Option<(Self::ProfileMetric, f64)>,
-        ) -> (Self::ProfileMetric, String) {
-            (0, "0 ms".to_string())
-        }
-
-        fn filter_llir_candidate(
-            &mut self,
-            _: &LLIRGraph,
-            _: CandidateFilterContext<'_>,
-        ) -> CandidateFilterResult {
-            if self.reject_candidates {
-                CandidateFilterResult::reject_with_display("test filter rejected candidate")
-            } else {
-                CandidateFilterResult::accept()
-            }
-        }
-    }
-
-    static PROFILE_CALLS: AtomicUsize = AtomicUsize::new(0);
-
-    #[derive(Default)]
-    struct CountingRuntime {
-        reject_candidates: bool,
-    }
-
-    impl Runtime for CountingRuntime {
-        type Ops = ();
-        type CompileArg = ();
-        type ExecReturn = ();
-        type ProfileMetric = usize;
-
-        fn initialize(_: Self::CompileArg) -> Self {
-            Self::default()
-        }
-
-        fn load_llir(&mut self, _: &LLIRGraph) {}
-
-        fn execute(&mut self, _: &DynMap) -> Self::ExecReturn {}
-
-        fn profile(
-            &mut self,
-            _: &LLIRGraph,
-            _: &DynMap,
-            _: usize,
-            _: Option<std::time::Duration>,
-            _: Option<(Self::ProfileMetric, f64)>,
-        ) -> (Self::ProfileMetric, String) {
-            let count = PROFILE_CALLS.fetch_add(1, Ordering::SeqCst);
-            (count, format!("{count} ms"))
-        }
-
-        fn filter_llir_candidate(
-            &mut self,
-            _: &LLIRGraph,
-            _: CandidateFilterContext<'_>,
-        ) -> CandidateFilterResult {
-            if self.reject_candidates {
-                CandidateFilterResult::reject_with_display("test filter rejected candidate")
-            } else {
-                CandidateFilterResult::accept()
-            }
-        }
-    }
-
-    #[derive(Default)]
-    struct PreparedCandidateRuntime {
-        prepared: usize,
-        prepared_profiles: usize,
-        fallback_profiles: usize,
-    }
-
-    impl Runtime for PreparedCandidateRuntime {
-        type Ops = ();
-        type CompileArg = ();
-        type ExecReturn = ();
-        type ProfileMetric = usize;
-
-        fn initialize(_: Self::CompileArg) -> Self {
-            Self::default()
-        }
-
-        fn load_llir(&mut self, _: &LLIRGraph) {}
-
-        fn execute(&mut self, _: &DynMap) -> Self::ExecReturn {}
-
-        fn profile(
-            &mut self,
-            _: &LLIRGraph,
-            _: &DynMap,
-            _: usize,
-            _: Option<std::time::Duration>,
-            _: Option<(Self::ProfileMetric, f64)>,
-        ) -> (Self::ProfileMetric, String) {
-            self.fallback_profiles += 1;
-            (0, "fallback".to_string())
-        }
-
-        fn prepare_profile_candidate(
-            &mut self,
-            _: &LLIRGraph,
-            _: CandidateFilterContext<'_>,
-        ) -> CandidateFilterResult {
-            self.prepared += 1;
-            CandidateFilterResult::accept()
-        }
-
-        fn profile_prepared_candidate(
-            &mut self,
-            _: &LLIRGraph,
-            _: &DynMap,
-            _: usize,
-            _: Option<std::time::Duration>,
-            _: Option<(Self::ProfileMetric, f64)>,
-            _: Option<ProfileBucketContext<'_>>,
-        ) -> (Self::ProfileMetric, String) {
-            self.prepared_profiles += 1;
-            (0, "prepared".to_string())
-        }
-    }
-
-    macro_rules! final_filter_test_op {
-        ($name:ident, $sort_name:literal, $rule_name:literal, $extracted:ty) => {
-            #[derive(Debug, Default)]
-            struct $name;
-
-            impl EgglogOp for $name {
-                fn sort(&self) -> SortDef {
-                    sort(OP_KIND, $sort_name, &[])
-                }
-
-                fn rewrites(&self) -> Vec<Rule> {
-                    vec![Rule::raw(format!(
-                        "(rule
-                            (
-                                (= ?sin (Op (Sin ?shape ?strides ?out_strides) ?inputs))
-                                (= (F32) (dtype ?sin))
-                            )
-                            (
-                                (let ?candidate (Op ({}) ?inputs))
-                                (union ?sin ?candidate)
-                                (set (dtype ?candidate) (F32))
-                            )
-                            :ruleset kernel_lower
-                            :name \"{}\"
-                        )",
-                        $sort_name, $rule_name
-                    ))]
-                }
-
-                fn cleanup(&self) -> bool {
-                    false
-                }
-
-                fn n_inputs(&self) -> usize {
-                    1
-                }
-
-                fn extract<'a>(
-                    &'a self,
-                    _: &'a SerializedEGraph,
-                    _: &[&'a ENodeId],
-                    input_enodes: Vec<&'a ENodeId>,
-                    _: &mut FxHashMap<&'a ENodeId, Vec<Expression>>,
-                    _: &mut FxHashMap<&'a ENodeId, Expression>,
-                ) -> (LLIROp, Vec<&'a ENodeId>) {
-                    (
-                        LLIROp::new::<dyn ReferenceOp>(
-                            Box::new(<$extracted>::default()) as Box<dyn ReferenceOp>
-                        ),
-                        input_enodes,
-                    )
-                }
-            }
-
-            impl ReferenceOp for $name {
-                fn execute(&self, inputs: Vec<&ReferenceData>, _: &DynMap) -> ReferenceData {
-                    let ReferenceData::F32(input) = inputs[0] else {
-                        panic!("final-filter test Sin candidates only support F32")
-                    };
-                    ReferenceData::F32(input.iter().map(|value| value.sin()).collect())
-                }
-            }
-        };
-    }
-
-    final_filter_test_op!(
-        FastButInvalidAfterUnroll,
-        "TestFastSin",
-        "test fast sin",
-        FastButInvalidAfterUnroll
-    );
-    final_filter_test_op!(
-        SlowerValidAfterUnroll,
-        "TestSafeSin",
-        "test safe sin",
-        SlowerValidAfterUnroll
-    );
-    final_filter_test_op!(
-        DuplicateFastLlir,
-        "DuplicateFastLlir",
-        "duplicate fast llir",
-        FastButInvalidAfterUnroll
-    );
-
-    #[derive(Default)]
-    struct DuplicateLlirRuntime {
-        profile_calls: usize,
-    }
-
-    impl Runtime for DuplicateLlirRuntime {
-        type Ops = (FastButInvalidAfterUnroll, DuplicateFastLlir);
-        type CompileArg = ();
-        type ExecReturn = ();
-        type ProfileMetric = usize;
-
-        fn initialize(_: Self::CompileArg) -> Self {
-            Self::default()
-        }
-
-        fn load_llir(&mut self, _: &LLIRGraph) {}
-
-        fn execute(&mut self, _: &DynMap) -> Self::ExecReturn {}
-
-        fn profile(
-            &mut self,
-            _: &LLIRGraph,
-            _: &DynMap,
-            _: usize,
-            _: Option<std::time::Duration>,
-            _: Option<(Self::ProfileMetric, f64)>,
-        ) -> (Self::ProfileMetric, String) {
-            self.profile_calls += 1;
-            (0, "same LLIR".to_string())
-        }
-    }
-
-    #[derive(Default)]
-    struct FinalFilterRuntime {
-        accepted_signatures: FxHashSet<(usize, usize, usize)>,
-        loaded_signatures: Vec<(usize, usize, usize)>,
-        profiled_fast: usize,
-        profiled_safe: usize,
-        rejected_unrolled_fast: usize,
-    }
-
-    impl FinalFilterRuntime {
-        fn signature(llir: &LLIRGraph) -> (usize, usize, usize) {
-            let mut fast = 0;
-            let mut safe = 0;
-            for op in llir.node_weights() {
-                let Some(reference_op) = op.to_dialect::<dyn ReferenceOp>() else {
-                    continue;
-                };
-                let reference_op = reference_op.as_ref().as_ref();
-                fast += usize::from(reference_op.as_any().is::<FastButInvalidAfterUnroll>());
-                safe += usize::from(reference_op.as_any().is::<SlowerValidAfterUnroll>());
-            }
-            (fast, safe, llir.node_count())
-        }
-    }
-
-    impl Runtime for FinalFilterRuntime {
-        type Ops = (FastButInvalidAfterUnroll, SlowerValidAfterUnroll);
-        type CompileArg = ();
-        type ExecReturn = ();
-        type ProfileMetric = usize;
-
-        fn initialize(_: Self::CompileArg) -> Self {
-            Self::default()
-        }
-
-        fn load_llir(&mut self, llir: &LLIRGraph) {
-            let signature = Self::signature(llir);
-            assert!(
-                self.accepted_signatures.contains(&signature),
-                "loaded final LLIR {signature:?} did not pass the hard candidate filter"
-            );
-            self.loaded_signatures.push(signature);
-        }
-
-        fn execute(&mut self, _: &DynMap) -> Self::ExecReturn {}
-
-        fn profile(
-            &mut self,
-            llir: &LLIRGraph,
-            _: &DynMap,
-            _: usize,
-            _: Option<std::time::Duration>,
-            _: Option<(Self::ProfileMetric, f64)>,
-        ) -> (Self::ProfileMetric, String) {
-            let (fast, safe, _) = Self::signature(llir);
-            assert_eq!(
-                fast + safe,
-                3,
-                "profiling should see the fully unrolled candidate"
-            );
-            if fast == 1 {
-                self.profiled_fast += 1;
-                (0, "fast".to_string())
-            } else {
-                self.profiled_safe += 1;
-                (1, "safe".to_string())
-            }
-        }
-
-        fn filter_llir_candidate(
-            &mut self,
-            llir: &LLIRGraph,
-            _: CandidateFilterContext<'_>,
-        ) -> CandidateFilterResult {
-            let signature = Self::signature(llir);
-            if signature.0 > 1 {
-                self.rejected_unrolled_fast += 1;
-                CandidateFilterResult::reject_with_display(
-                    "fast candidate exceeds final unrolled resource limit",
-                )
-            } else {
-                self.accepted_signatures.insert(signature);
-                CandidateFilterResult::accept()
-            }
-        }
-    }
-
-    #[derive(Default)]
-    struct ProfileDimFinalFilterRuntime {
-        last_profile_dim: Option<usize>,
-        loaded_signatures: Vec<(usize, usize, usize)>,
-        rejected_unrolled_fast: usize,
-    }
-
-    impl Runtime for ProfileDimFinalFilterRuntime {
-        type Ops = (FastButInvalidAfterUnroll, SlowerValidAfterUnroll);
-        type CompileArg = ();
-        type ExecReturn = ();
-        type ProfileMetric = usize;
-
-        fn initialize(_: Self::CompileArg) -> Self {
-            Self::default()
-        }
-
-        fn load_llir(&mut self, llir: &LLIRGraph) {
-            let signature = FinalFilterRuntime::signature(llir);
-            let profile_dim = self
-                .last_profile_dim
-                .expect("a selected LLIR must be loaded after profiling");
-            assert!(
-                signature.0 == 0 || signature.0 + signature.1 == 1 || profile_dim <= 1,
-                "load rejected unrolled fast LLIR {signature:?} at profile dim {profile_dim}"
-            );
-            self.loaded_signatures.push(signature);
-        }
-
-        fn execute(&mut self, _: &DynMap) -> Self::ExecReturn {}
-
-        fn profile(
-            &mut self,
-            llir: &LLIRGraph,
-            dyn_map: &DynMap,
-            _: usize,
-            _: Option<std::time::Duration>,
-            _: Option<(Self::ProfileMetric, f64)>,
-        ) -> (Self::ProfileMetric, String) {
-            let (fast, safe, _) = FinalFilterRuntime::signature(llir);
-            assert_eq!(
-                fast + safe,
-                3,
-                "profiling should see the fully unrolled candidate"
-            );
-            self.last_profile_dim = dyn_map.get(&Symbol::from('s')).copied();
-            if fast == 1 {
-                (0, "fast".to_string())
-            } else {
-                (1, "safe".to_string())
-            }
-        }
-
-        fn filter_llir_candidate(
-            &mut self,
-            llir: &LLIRGraph,
-            context: CandidateFilterContext<'_>,
-        ) -> CandidateFilterResult {
-            let (fast, safe, _) = FinalFilterRuntime::signature(llir);
-            let is_unrolled = fast + safe > 1;
-            let filter_dim = context
-                .dyn_map
-                .get(&Symbol::from('s'))
-                .copied()
-                .unwrap_or(1);
-            if is_unrolled && fast > 0 && filter_dim > 1 {
-                self.rejected_unrolled_fast += 1;
-                CandidateFilterResult::reject_with_display(format!(
-                    "fast candidate exceeds final resource limit at s={filter_dim}"
-                ))
-            } else {
-                CandidateFilterResult::accept()
-            }
-        }
-    }
-
-    #[derive(Default)]
-    struct AggregateBucketFilterRuntime {
-        individually_accepted: FxHashSet<(usize, usize, usize)>,
-        aggregate_attempts: Vec<Vec<(usize, usize, usize)>>,
-        aggregate_accepted: FxHashSet<Vec<(usize, usize, usize)>>,
-        loaded_sets: Vec<Vec<(usize, usize, usize)>>,
-    }
-
-    impl Runtime for AggregateBucketFilterRuntime {
-        type Ops = (FastButInvalidAfterUnroll, SlowerValidAfterUnroll);
-        type CompileArg = ();
-        type ExecReturn = ();
-        type ProfileMetric = usize;
-
-        fn initialize(_: Self::CompileArg) -> Self {
-            Self::default()
-        }
-
-        fn load_llir(&mut self, _: &LLIRGraph) {
-            panic!("aggregate bucket regression must load a retained bucket set")
-        }
-
-        fn load_llir_buckets(
-            &mut self,
-            _: &FxHashMap<Symbol, Vec<DimBucket>>,
-            bucket_llirs: &[BucketLLIR],
-        ) {
-            let signatures = bucket_llirs
-                .iter()
-                .map(|(_, _, llir)| FinalFilterRuntime::signature(llir))
-                .collect_vec();
-            assert!(
-                signatures
-                    .iter()
-                    .all(|signature| self.individually_accepted.contains(signature)),
-                "load received a bucket LLIR that did not pass individual final filtering"
-            );
-            assert!(
-                self.aggregate_accepted.contains(&signatures),
-                "load received bucket set {signatures:?} that did not pass aggregate filtering"
-            );
-            self.loaded_sets.push(signatures);
-        }
-
-        fn execute(&mut self, _: &DynMap) -> Self::ExecReturn {}
-
-        fn profile(
-            &mut self,
-            llir: &LLIRGraph,
-            _: &DynMap,
-            _: usize,
-            _: Option<std::time::Duration>,
-            _: Option<(Self::ProfileMetric, f64)>,
-        ) -> (Self::ProfileMetric, String) {
-            let (fast, safe, _) = FinalFilterRuntime::signature(llir);
-            if fast > 0 {
-                (0, "fast".to_string())
-            } else if safe > 0 {
-                (1, "safe".to_string())
-            } else {
-                (2, "original".to_string())
-            }
-        }
-
-        fn aggregate_profile_metrics(metrics: &[Self::ProfileMetric]) -> Self::ProfileMetric {
-            metrics.iter().sum()
-        }
-
-        fn filter_llir_candidate(
-            &mut self,
-            llir: &LLIRGraph,
-            _: CandidateFilterContext<'_>,
-        ) -> CandidateFilterResult {
-            self.individually_accepted
-                .insert(FinalFilterRuntime::signature(llir));
-            CandidateFilterResult::accept()
-        }
-
-        fn filter_llir_bucket_set(
-            &mut self,
-            _: &FxHashMap<Symbol, Vec<DimBucket>>,
-            bucket_llirs: &[BucketLLIRRef<'_>],
-            _: &CompileOptions,
-        ) -> CandidateFilterResult {
-            let signatures = bucket_llirs
-                .iter()
-                .map(|bucket| FinalFilterRuntime::signature(bucket.llir))
-                .collect_vec();
-            assert!(
-                signatures
-                    .iter()
-                    .all(|signature| self.individually_accepted.contains(signature)),
-                "aggregate filter received a bucket that skipped individual filtering"
-            );
-            self.aggregate_attempts.push(signatures.clone());
-            let all_fast = signatures
-                .iter()
-                .all(|(fast, safe, _)| *fast > 0 && *safe == 0);
-            if all_fast {
-                CandidateFilterResult::reject_with_display(
-                    "fast bucket finalists conflict in retained resources",
-                )
-            } else {
-                self.aggregate_accepted.insert(signatures);
-                CandidateFilterResult::accept()
-            }
-        }
-    }
-
-    static SEARCH_BUDGET_PROFILE_CALLS: AtomicUsize = AtomicUsize::new(0);
-    static SEARCH_BUDGET_FINAL_FILTER_CALLS: AtomicUsize = AtomicUsize::new(0);
-
-    #[derive(Debug, Default)]
-    struct SearchBudgetFinalFilterRuntime;
-
-    impl Runtime for SearchBudgetFinalFilterRuntime {
-        type Ops = (FastButInvalidAfterUnroll, SlowerValidAfterUnroll);
-        type CompileArg = ();
-        type ExecReturn = ();
-        type ProfileMetric = usize;
-
-        fn initialize(_: Self::CompileArg) -> Self {
-            Self
-        }
-
-        fn load_llir(&mut self, _: &LLIRGraph) {
-            panic!("a runtime with no viable final candidate must not be loaded")
-        }
-
-        fn execute(&mut self, _: &DynMap) -> Self::ExecReturn {}
-
-        fn profile(
-            &mut self,
-            llir: &LLIRGraph,
-            _: &DynMap,
-            _: usize,
-            _: Option<std::time::Duration>,
-            _: Option<(Self::ProfileMetric, f64)>,
-        ) -> (Self::ProfileMetric, String) {
-            let profile_call = SEARCH_BUDGET_PROFILE_CALLS.fetch_add(1, Ordering::SeqCst);
-            if profile_call == 1 {
-                // Cross the global budget only after a second genome has
-                // entered profiling, so it is retained as a possible fallback.
-                std::thread::sleep(std::time::Duration::from_millis(150));
-            }
-            let (fast, _, _) = FinalFilterRuntime::signature(llir);
-            if fast == 1 {
-                (0, "fast".to_string())
-            } else {
-                (1, "safe".to_string())
-            }
-        }
-
-        fn filter_llir_candidate(
-            &mut self,
-            llir: &LLIRGraph,
-            _: CandidateFilterContext<'_>,
-        ) -> CandidateFilterResult {
-            let (fast, safe, _) = FinalFilterRuntime::signature(llir);
-            if fast + safe > 1 {
-                // Candidates are filtered unrolled at search time AND at
-                // finalization now. Accept the first two calls (the two
-                // search candidates), then reject the finalization
-                // re-checks, burning the search time limit on the first.
-                let call = SEARCH_BUDGET_FINAL_FILTER_CALLS.fetch_add(1, Ordering::SeqCst);
-                if call < 2 {
-                    return CandidateFilterResult::accept();
-                }
-                std::thread::sleep(std::time::Duration::from_millis(120));
-                CandidateFilterResult::reject_with_display("forced final rejection")
-            } else {
-                CandidateFilterResult::accept()
-            }
-        }
-    }
-
-    static CANDIDATE_BUDGET_FINAL_FILTER_CALLS: AtomicUsize = AtomicUsize::new(0);
-
-    #[derive(Debug, Default)]
-    struct CandidateBudgetFinalFilterRuntime;
-
-    impl Runtime for CandidateBudgetFinalFilterRuntime {
-        type Ops = (FastButInvalidAfterUnroll, SlowerValidAfterUnroll);
-        type CompileArg = ();
-        type ExecReturn = ();
-        type ProfileMetric = usize;
-
-        fn initialize(_: Self::CompileArg) -> Self {
-            Self
-        }
-
-        fn load_llir(&mut self, _: &LLIRGraph) {
-            panic!("a timed-out final candidate must not be loaded")
-        }
-
-        fn execute(&mut self, _: &DynMap) -> Self::ExecReturn {}
-
-        fn profile(
-            &mut self,
-            llir: &LLIRGraph,
-            _: &DynMap,
-            _: usize,
-            _: Option<std::time::Duration>,
-            _: Option<(Self::ProfileMetric, f64)>,
-        ) -> (Self::ProfileMetric, String) {
-            let (fast, _, _) = FinalFilterRuntime::signature(llir);
-            if fast == 1 {
-                (0, "fast".to_string())
-            } else {
-                (1, "safe".to_string())
-            }
-        }
-
-        fn filter_llir_candidate(
-            &mut self,
-            llir: &LLIRGraph,
-            _: CandidateFilterContext<'_>,
-        ) -> CandidateFilterResult {
-            let (fast, safe, _) = FinalFilterRuntime::signature(llir);
-            if fast + safe > 1 {
-                CANDIDATE_BUDGET_FINAL_FILTER_CALLS.fetch_add(1, Ordering::SeqCst);
-                std::thread::sleep(std::time::Duration::from_millis(30));
-            }
-            CandidateFilterResult::accept()
-        }
-    }
-
-    #[test]
-    fn compile_options_defaults_and_search_time_limit_builder() {
-        let opts = CompileOptions::default();
-        assert_eq!(opts.limit, 100);
-        assert_eq!(opts.trials, 3);
-        assert_eq!(opts.search_time_limit, std::time::Duration::MAX);
-        assert!(!opts.egglog_log);
-        assert!(!opts.rolling_log);
-        assert!(opts.search_log);
-        assert!(opts.search_dims.is_empty());
-
-        let time_limit = std::time::Duration::from_millis(25);
-        let opts = CompileOptions::default()
-            .search_graph_limit(7)
-            .search_time_limit(time_limit)
-            .search_dim('c', 16)
-            .egglog_log(true)
-            .rolling_log(true)
-            .search_log(false);
-        assert_eq!(opts.limit, 7);
-        assert_eq!(opts.search_time_limit, time_limit);
-        assert_eq!(opts.search_dims[&sym("c")], 16);
-        assert!(opts.egglog_log);
-        assert!(opts.rolling_log);
-        assert!(!opts.search_log);
-    }
-
-    #[test]
-    fn compile_applies_search_dims_after_build_with_documented_precedence() {
-        SEARCH_DIM_LATE_PASS_CALLED.store(false, Ordering::SeqCst);
-        SEARCH_DIM_LATE_PASS_SAW_C.store(false, Ordering::SeqCst);
-
-        let mut cx = Graph::new();
-        let _ = cx.tensor(('s', 'c')).output();
-        let options = CompileOptions::default()
-            .dim_buckets('s', &[DimBucket::new(1, 4).representative(3)])
-            .search_dim('s', 4)
-            .search_dim('c', 16)
-            .profile_dim('s', 2)
-            .profile_dim('c', 7)
-            .search_graph_limit(1)
-            .search_log(false);
-
-        let runtime = cx.compile(SearchDimRecordingRuntime::default(), options);
-
-        assert!(SEARCH_DIM_LATE_PASS_CALLED.load(Ordering::SeqCst));
-        assert!(
-            !SEARCH_DIM_LATE_PASS_SAW_C.load(Ordering::SeqCst),
-            "search-only dimensions must not leak into build-time late passes"
-        );
-        assert_eq!(runtime.profile_dyn_maps.len(), 1);
-        assert_eq!(runtime.profile_dyn_maps[0][&sym("s")], 2);
-        assert_eq!(runtime.profile_dyn_maps[0][&sym("c")], 7);
-        assert_eq!(runtime.bucket_representative_dyn_maps.len(), 1);
-        assert_eq!(runtime.bucket_representative_dyn_maps[0][&sym("s")], 3);
-        assert_eq!(runtime.bucket_representative_dyn_maps[0][&sym("c")], 16);
-        assert_eq!(cx.dyn_map[&sym("s")], 4);
-        assert_eq!(cx.dyn_map[&sym("c")], 16);
-    }
-
-    #[test]
-    fn search_time_limit_stops_after_initial_viable_candidate() {
-        let mut cx = Graph::new();
-        let a = cx.tensor((4, 8));
-        let b = cx.tensor((8, 4));
-        let c = cx.tensor((4, 4));
-        let _ = (a.matmul(b) + c).relu().softmax(1).output();
-
-        cx.build_search_space::<CountingRuntime>(CompileOptions::default());
-
-        PROFILE_CALLS.store(0, Ordering::SeqCst);
-        let _ = cx.search(
-            CountingRuntime::default(),
-            CompileOptions::default().search_time_limit(std::time::Duration::ZERO),
-        );
-        assert_eq!(PROFILE_CALLS.load(Ordering::SeqCst), 1);
-    }
-
-    #[test]
-    fn build_search_space_does_not_configure_runtime_filter() {
-        let mut cx = Graph::new();
-        let _ = cx.tensor(1).output();
-
-        cx.build_search_space::<TestFilterRuntime>(CompileOptions::default());
-        assert_eq!(cx.egraphs.len(), 1);
-    }
-
-    #[test]
-    #[should_panic(expected = "runtime filter failures")]
-    fn runtime_filter_rejects_candidates() {
-        let mut cx = Graph::new();
-        let _ = cx.tensor(1).output();
-        cx.build_search_space::<TestFilterRuntime>(CompileOptions::default());
-
-        let _ = cx.search(
-            TestFilterRuntime {
-                reject_candidates: true,
-            },
-            CompileOptions::default().search_graph_limit(1),
-        );
-    }
-
-    #[test]
-    fn runtime_filter_rejects_candidate_before_profile() {
-        let mut cx = Graph::new();
-        let _ = cx.tensor(1).output();
-        cx.build_search_space::<CountingRuntime>(CompileOptions::default());
-
-        PROFILE_CALLS.store(0, Ordering::SeqCst);
-        let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-            let _ = cx.search(
-                CountingRuntime {
-                    reject_candidates: true,
-                },
-                CompileOptions::default().search_graph_limit(1),
-            );
-        }));
-
-        assert!(result.is_err());
-        assert_eq!(PROFILE_CALLS.load(Ordering::SeqCst), 0);
-    }
-
-    #[test]
-    fn search_profiles_the_candidate_prepared_by_the_runtime() {
-        let mut cx = Graph::new();
-        let _ = cx.tensor(1).output();
-
-        let runtime = cx.compile(
-            PreparedCandidateRuntime::default(),
-            CompileOptions::default()
-                .search_graph_limit(1)
-                .search_log(false),
-        );
-
-        assert_eq!(runtime.prepared, 1);
-        assert_eq!(runtime.prepared_profiles, 1);
-        assert_eq!(
-            runtime.fallback_profiles, 0,
-            "a prepared candidate must not be loaded again through Runtime::profile"
-        );
-    }
-
-    #[test]
-    fn search_profiles_each_extracted_llir_only_once() {
-        let mut cx = Graph::new();
-        let input = cx.tensor(8);
-        let _ = input.sin().output();
-
-        let options = CompileOptions::default()
-            .search_graph_limit(16)
-            .generation_size(16)
-            .mutations(1)
-            .search_log(false);
-        let mut rng = rand::rngs::StdRng::seed_from_u64(0xDED0_0111);
-        let runtime = cx.compile_with_rng(DuplicateLlirRuntime::default(), options, &mut rng);
-
-        assert_eq!(
-            runtime.profile_calls, 1,
-            "two e-graph choices that extract to the same LLIR must consume one search slot"
-        );
-    }
-
-    #[test]
-    fn final_filter_skips_fast_invalid_unroll_and_loads_filtered_fallback() {
-        let mut cx = Graph::new();
-        let input = cx.tensor(8);
-        let _ = input.sin().sin().sin().output();
-
-        let options = CompileOptions::default()
-            .search_graph_limit(32)
-            .generation_size(8)
-            .mutations(2)
-            .search_log(false);
-        let mut rng = rand::rngs::StdRng::seed_from_u64(0xF1A1_F11E);
-        let runtime = cx.compile_with_rng(FinalFilterRuntime::default(), options, &mut rng);
-
-        assert_eq!(
-            runtime.profiled_fast, 0,
-            "invalid-unrolled fast candidates must be filtered before profiling"
-        );
-        assert!(runtime.profiled_safe > 0, "safe candidate was not profiled");
-        assert!(
-            runtime.rejected_unrolled_fast > 0,
-            "the fast candidate should be rejected by the unrolled candidate filter"
-        );
-        assert_eq!(
-            runtime.loaded_signatures.len(),
-            1,
-            "search should load exactly one final LLIR"
-        );
-        let (fast, safe, _) = runtime.loaded_signatures[0];
-        assert_eq!(fast, 0, "the rejected fast candidate was loaded");
-        assert!(
-            safe > 1,
-            "the loaded fallback should be the fully unrolled safe graph"
-        );
-    }
-
-    #[derive(Default)]
-    struct EarlyStopRecordingRuntime {
-        early_stop_args: Vec<Option<(usize, f64)>>,
-        profile_calls: usize,
-    }
-
-    impl Runtime for EarlyStopRecordingRuntime {
-        // Two competing ops so the e-graph has real choices and the search
-        // actually produces offspring after the initial genome.
-        type Ops = (FastButInvalidAfterUnroll, SlowerValidAfterUnroll);
-        type CompileArg = ();
-        type ExecReturn = ();
-        type ProfileMetric = usize;
-
-        fn initialize(_: Self::CompileArg) -> Self {
-            Self::default()
-        }
-
-        fn load_llir(&mut self, _: &LLIRGraph) {}
-
-        fn execute(&mut self, _: &DynMap) -> Self::ExecReturn {}
-
-        fn profile(
-            &mut self,
-            _: &LLIRGraph,
-            _: &DynMap,
-            _: usize,
-            _: Option<std::time::Duration>,
-            early_stop: Option<(Self::ProfileMetric, f64)>,
-        ) -> (Self::ProfileMetric, String) {
-            self.early_stop_args.push(early_stop);
-            // Strictly increasing metrics: the initial genome (metric 0)
-            // stays the running best for the whole search.
-            let metric = self.profile_calls;
-            self.profile_calls += 1;
-            (metric, format!("{metric} ms"))
-        }
-    }
-
-    #[test]
-    fn search_passes_best_so_far_to_profile_early_stop() {
-        let mut cx = Graph::new();
-        let input = cx.tensor(8);
-        let _ = input.sin().sin().sin().output();
-
-        let options = CompileOptions::default()
-            .search_graph_limit(16)
-            .generation_size(4)
-            .mutations(2)
-            .early_stop_factor(2.0)
-            .search_log(false);
-        let mut rng = rand::rngs::StdRng::seed_from_u64(0xEA51_5709);
-        let runtime = cx.compile_with_rng(EarlyStopRecordingRuntime::default(), options, &mut rng);
-
-        assert!(
-            runtime.early_stop_args.len() > 1,
-            "search should profile more candidates than the initial genome"
-        );
-        assert_eq!(
-            runtime.early_stop_args[0], None,
-            "no best exists before the initial genome is profiled"
-        );
-        for arg in &runtime.early_stop_args[1..] {
-            let (best, factor) =
-                arg.expect("offspring profiling should receive the best-so-far cutoff");
-            assert_eq!(best, 0, "the initial genome's metric is the running best");
-            assert_eq!(factor, 2.0);
-        }
-    }
-
-    #[test]
-    fn unbucketed_final_filter_uses_profile_dims_seen_by_load() {
-        let mut cx = Graph::new();
-        cx.set_dim('s', 1);
-        let input = cx.tensor('s');
-        let _ = input.sin().sin().sin().output();
-
-        let options = CompileOptions::default()
-            .profile_dim('s', 8)
-            .search_graph_limit(32)
-            .generation_size(8)
-            .mutations(2)
-            .search_log(false);
-        let mut rng = rand::rngs::StdRng::seed_from_u64(0xF1A1_F11E);
-        let runtime =
-            cx.compile_with_rng(ProfileDimFinalFilterRuntime::default(), options, &mut rng);
-
-        assert_eq!(runtime.last_profile_dim, Some(8));
-        assert!(
-            runtime.rejected_unrolled_fast > 0,
-            "the final hard filter did not receive the profiling dimension"
-        );
-        assert_eq!(runtime.loaded_signatures.len(), 1);
-        let (fast, safe, _) = runtime.loaded_signatures[0];
-        assert_eq!(fast, 0, "load received a candidate invalid at profile_dim");
-        assert!(safe > 1, "load should receive a fully unrolled fallback");
-    }
-
-    #[test]
-    fn final_fallback_validates_one_candidate_then_stops_at_search_time_limit() {
-        SEARCH_BUDGET_PROFILE_CALLS.store(0, Ordering::SeqCst);
-        SEARCH_BUDGET_FINAL_FILTER_CALLS.store(0, Ordering::SeqCst);
-
-        let mut cx = Graph::new();
-        let input = cx.tensor(8);
-        let _ = input.sin().sin().sin().output();
-        let options = CompileOptions::default()
-            .search_graph_limit(2)
-            .generation_size(1)
-            .mutations(2)
-            .search_time_limit(std::time::Duration::from_millis(100))
-            .search_log(false);
-        let mut rng = rand::rngs::StdRng::seed_from_u64(0xF1A1_F11E);
-        let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-            cx.compile_with_rng(SearchBudgetFinalFilterRuntime, options, &mut rng)
-        }));
-
-        let panic = result.expect_err("every finalized candidate should be rejected");
-        let message = panic
-            .downcast_ref::<String>()
-            .map(String::as_str)
-            .or_else(|| panic.downcast_ref::<&str>().copied())
-            .unwrap_or("non-string panic");
-        assert!(
-            message.contains("search time limit expired before finalizing ranked candidate 2"),
-            "unexpected finalization failure: {message}"
-        );
-        assert_eq!(
-            SEARCH_BUDGET_PROFILE_CALLS.load(Ordering::SeqCst),
-            2,
-            "the fallback genome must be retained before the budget expires"
-        );
-        assert_eq!(
-            SEARCH_BUDGET_FINAL_FILTER_CALLS.load(Ordering::SeqCst),
-            3,
-            "two search-time filter calls plus one finalization attempt; no fallback after expiry"
-        );
-    }
-
-    #[test]
-    fn final_fallback_rejects_candidates_that_exceed_candidate_timeout() {
-        CANDIDATE_BUDGET_FINAL_FILTER_CALLS.store(0, Ordering::SeqCst);
-
-        let mut cx = Graph::new();
-        let input = cx.tensor(8);
-        let _ = input.sin().sin().sin().output();
-        let options = CompileOptions::default()
-            .search_graph_limit(2)
-            .generation_size(1)
-            .mutations(2)
-            .candidate_timeout(std::time::Duration::from_millis(20))
-            .search_log(false);
-        let mut rng = rand::rngs::StdRng::seed_from_u64(0xF1A1_F11E);
-        let result = std::panic::catch_unwind(std::panic::AssertUnwindSafe(|| {
-            cx.compile_with_rng(CandidateBudgetFinalFilterRuntime, options, &mut rng)
-        }));
-
-        let panic = result.expect_err("timed-out finalists must not be loaded");
-        let message = panic
-            .downcast_ref::<String>()
-            .map(String::as_str)
-            .or_else(|| panic.downcast_ref::<&str>().copied())
-            .unwrap_or("non-string panic");
-        assert!(
-            message.contains("candidate timeout expired while finalizing ranked candidate 2"),
-            "unexpected finalization failure: {message}"
-        );
-        assert_eq!(
-            CANDIDATE_BUDGET_FINAL_FILTER_CALLS.load(Ordering::SeqCst),
-            4,
-            "both candidates are filtered unrolled at search time and again at finalization"
-        );
-    }
-
-    #[test]
-    fn aggregate_bucket_filter_backtracks_to_fastest_viable_retained_set() {
-        let mut cx = Graph::new();
-        let _ = cx.tensor('s').sin().output();
-
-        let options = CompileOptions::default()
-            .dim_buckets('s', &[DimBucket::new(1, 1), DimBucket::new(2, 2)])
-            .search_graph_limit(32)
-            .generation_size(8)
-            .mutations(2)
-            .search_log(false);
-        let mut rng = rand::rngs::StdRng::seed_from_u64(0xA66E_6A7E);
-        let runtime =
-            cx.compile_with_rng(AggregateBucketFilterRuntime::default(), options, &mut rng);
-
-        assert!(
-            runtime.aggregate_attempts.len() >= 2,
-            "aggregate rejection should trigger a slower finalist combination"
-        );
-        assert!(
-            runtime.aggregate_attempts[0]
-                .iter()
-                .all(|(fast, safe, _)| *fast > 0 && *safe == 0),
-            "the independently fastest all-fast set should be tried first"
-        );
-        assert_eq!(runtime.loaded_sets.len(), 1);
-        let loaded = &runtime.loaded_sets[0];
-        assert_eq!(
-            loaded
-                .iter()
-                .filter(|(fast, safe, _)| *fast > 0 && *safe == 0)
-                .count(),
-            1,
-            "best-first fallback should keep one fast bucket"
-        );
-        assert_eq!(
-            loaded.iter().filter(|(_, safe, _)| *safe > 0).count(),
-            1,
-            "best-first fallback should replace only one conflicting finalist"
-        );
-    }
-
-    #[test]
-    fn compile_builds_search_space_and_searches_it() {
-        let mut cx = Graph::new();
-        let _ = cx.tensor(1).output();
-
-        let _ = cx.compile(
-            TestFilterRuntime::default(),
-            CompileOptions::default().search_graph_limit(1),
-        );
-
-        assert_eq!(cx.egraphs.len(), 1);
-    }
-
-    #[test]
-    fn bucketed_build_search_space_builds_one_egraph_per_bucket() {
-        let mut cx = Graph::new();
-        let _ = cx.tensor('s').output();
-
-        cx.build_search_space::<TestFilterRuntime>(
-            CompileOptions::default()
-                .dim_buckets('s', &[DimBucket::new(1, 1), DimBucket::new(2, 4)]),
-        );
-
-        assert_eq!(cx.egraphs.len(), 2);
-        assert_eq!(cx.egraph_contexts.len(), 2);
-        assert_eq!(
-            cx.egraph_contexts[0].intervals[&sym("s")],
-            DimInterval::new(1, 1)
-        );
-        assert_eq!(
-            cx.egraph_contexts[1].intervals[&sym("s")],
-            DimInterval::new(2, 4)
-        );
-    }
-
-    #[test]
-    #[should_panic(expected = "search cannot change buckets after build")]
-    fn search_with_rng_dim_buckets_after_build_search_space_panics() {
-        let mut cx = Graph::new();
-        let _ = cx.tensor('s').output();
-
-        cx.build_search_space::<TestFilterRuntime>(CompileOptions::default());
-        let mut rng = rand::rng();
-        let _ = cx.search_with_rng(
-            TestFilterRuntime::default(),
-            CompileOptions::default().search_graph_limit(1).dim_buckets(
-                's',
-                &[DimBucket::new(1, 1), DimBucket::new(2, 4).representative(4)],
-            ),
-            &mut rng,
         );
     }
